Reads MCMC estimates from Awatea (Paul Starr) processed r file structure into an operating model

A function that uses the file location of a fitted Awatea model post-processed into a set of rmd files

Awatea2OM(AwateaDir, nsim = 48, proyears = 50, Name = NULL,
  Source = "No source provided", Author = "No author provided",
  verbose = T)

Arguments

AwateaDir
A folder with Awatea files

nsim
The number of simulations to take for parameters with uncertainty (for OM@cpars custom parameters)

proyears
The number of projection years for MSE

Name
The name of the operating model

Source
Reference to assessment documentation e.g. a url

Author
Who did the assessment

verbose
Should the r4ss function SS_ouput return detailed messages?