Check Convergence

Have I undertaken enough simulations (nsim)? Has my MSE converged on stable (reliable) peformance metrics?

Converge(MSEobj, PMs = list(Yield, P10, AAVY), maxMP = 15,
  thresh = 0.5, ref.it = 20, inc.leg = FALSE, all.its = FALSE,
  nrow = NULL, ncol = NULL)

Arguments

MSEobj
An MSE object of class 'MSE'

PMs
A list of PM objects

maxMP
Maximum number of MPs to include in a single plot

thresh
The convergence threshold. Maximum root mean square deviation over the last ref.it iterations

ref.it
The number of iterations to calculate the convergence statistics. For example, a value of 20 means convergence diagnostics are calculated over last 20 simulations

inc.leg
Logical. Should the legend be displayed?

all.its
Logical. Plot all iterations? Otherwise only (nsim-ref.it):nsim

nrow
Numeric. Optional. Number of rows

ncol
Numeric. Optional. Number of columns

Details

Performance metrics are plotted against the number of simulations. Convergence diagonostics are calculated over the last ref.it (default = 20) iterations. The convergence diagnostics are:

Is the order of the MPs stable over the last ref.it iterations?

Is the average difference in performance statistic over the last ref.it iterations < thresh?

By default three commonly used performance metrics are used:

Average Yield Relative to Reference Yield

Probability Spawning Biomass is above 0.1BMSY

Probability Average Annual Variability in Yield is < 20 per cent

Additional or alternative performance metrics objects can be supplied. Advanced users can develop their own performance metrics.

Note

See relevant section of the DLMtool User Guide for more information.

Examples

# NOT RUN {
MSE <- runMSE()
Converge(MSE)
# }