Depletion-Based Stock Reduction Analysis

User prescribed BMSY/B0, M, FMSY/M are used to find B0 and therefore the OFL by back-constructing the stock to match a user specified level of stock depletion (OFL = M * FMSY/M * depletion* B0).

DBSRA(x, Data, reps = 100)

DBSRA_40(x, Data, reps = 100)

DBSRA4010(x, Data, reps = 100)

Arguments

x
A position in a data-limited methods object.

Data
A data-limited methods object.

reps
The number of samples of the TAC (OFL) recommendation.

Value

A Rec object with a vector of TAC (OFL) values.

Functions

DBSRA: Base version. You specify a range of stock depletion and, given historical catches DB-SRA calculates what unfished biomass must have been to get you here given samples for M, FMSY relative to M and also BMSY relative to Bunfished.

DBSRA_40: Assumes 40 percent current depletion (Bcurrent/B0 = 0.4), which is more or less the most optimistic state for a stock (ie very close to BMSY/B0 for many stocks).

DBSRA4010: Base version paired with the 40-10 rule that throttles back the OFL to zero at 10 percent of unfished biomass.

Note

This is set up to return the OFL (FMSY * current biomass).

You may have noticed that you -the user- specify three of the factors that make the quota recommendation. So this can be quite a subjective method.

Also the DB-SRA method of this package isn't exactly the same as the original method of Dick and MacCall (2011) because it has to work for simulated depletions above BMSY/B0 and even on occasion over B0. Also it doesn't have the modification for flatfish life histories that has previously been applied by Dick and MacCall.

References

Dick, E.J., MacCall, A.D., 2010. Estimates of sustainable yield for 50 data-poor stocks in the Pacific Coast groundfish fishery management plan. Technical memorandum. Southwest fisheries Science Centre, Santa Cruz, CA. National Marine Fisheries Service, National Oceanic and Atmospheric Administration of the U.S. Department of Commerce. NOAA-TM-NMFS-SWFSC-460.

Dick, E.J., MacCall, A.D., 2011. Depletion-Based Stock Reduction Analysis: A catch-based method for determining sustainable yields for data-poor fish stocks. Fish. Res. 110, 331-341.