Data-Limited Methods Toolkit

Simulation testing and implementation of data-limited fishery stock assessment methods

Additional Information

See the DLMtool User Guide for a detailed description of how to use the DLMtool package.

The help documentation for the DLMtool package can also be accessed here.

See the Data-Limited Toolkit Website for more information on the DLMtool, including an interactive demo of the main features of the toolkit, information on case studies where the toolkit has been applied, and more about the history and development of the DLMtool.

References

Carruthers, T.R., Punt, A.E., Walters, C.J., MacCall, A., McAllister, M.K., Dick, E.J., Cope, J. 2014. Evaluating methods for setting catch limits in data-limited fisheries. Fisheries Research. 153: 48-68.

Carruthers, T.R., Kell, L.T., Butterworth, D.S., Maunder, M.N., Geromont, H.F., Walters, C., McAllister, M.K., Hillary, R., Levontin, P., Kitakado, T., Davies, C.R. Performance review of simple management procedures. ICES Journal of Marine Science.