Class 'Hist'

An object for storing information generated by the end of the historical simulations

Slots

Data

The Data object at the end of the historical period

Obs

A table of sampled parameters of the observation model. Table of nsim rows. Real numbers

Abias: bias in observed current absolute stock biomass

Aerr: error in observed current absolute stock biomass

betas: hyper-stability/hyper-depletion parameter

BMSY_B0bias: bias in ratio of most productive stock size relative to unfished

Brefbias: bias in BMSY stock levels (target or reference biomass levels)

CAA_ESS: the effective sample size of multinomial catch-at-age observation model (number of independent draws)

CAA_nsamp: the number of catch-at-age observations per time step

CAL_ESS: the effective sample size of multinomial catch-at-length observation model (number of independent draws)

CAL_nsamp: the number of catch-at-length observations per time step

Cbias: bias in observed catches

Crefbias: bias in MSY prediction (target or reference catch)

Csd: observation error in observed catches (lognormal CV)

Dbias: bias in observed stock depletion (also applies to depletion Dt for DCAC)

Derr: error in observed stock depletion

FMSY_Mbias: bias in ratio of FMSY to natural mortality rate

hbias: bias in observed steepness of the stock recruitment relationship

Irefbias: bias in abundance index corresponding to BMSY stock levels

Isd: observation error in relative abundance index (lognormal CV)

Kbias: bias in maximum growth rate (von Bertalanffy K parameter)

lenMbias: bias in length at 50 per cent maturity

LFCbias: bias in length at first capture

LFSbias: bias in length at full selection

Linfbias: bias in maximum length (von Bertalanffy Linf parameter)

Mbias: bias in observed natural mortality rate

Recsd: error in observed recruitment

t0bias: bias in theoretical length at age zero (von Bertalanffy t0 parameter)

OM

A numeric data.frame with nsim rows with sampled Stock & Fleet parameters

AtAge

A named list with arrays (dim nsim, maxage, nyears+proyears):

Length: Length-at-age for each simulation, age, and year

Weight: Weight-at-age for each simulation, age, and year

Select: Selectivity-at-age for each simulation, age, and year

Retention: Retention-at-age for each simulation, age, and year

Maturity: Maturity-at-age for each simulation, age, and year

N.Mortality: Natural mortality-at-age for each simulation, age, and year

Nage: Total numbers by simulation, age, and year

SSBage: Spawning stock biomass by simulation, age, and year

FM: Fishing mortality by simulation, age, year, and area

TSdata

A named list with population dynamics by simulation and year :

VB: Vulnerable biomass

SSB: Spawning stock biomass

B: Total biomass

Removals: Removals

Catch: Retained catch (will be same as removals unless there is discard mortality)

Rec: Recruitment

N: Total numbers

Find: Historical fishing effort

Marray: Average adult natural mortality (historical & projection)

RecDev: Recruitment deviations (historical & projection)

Ref

A numeric data.frame with nsim rows containing biological reference points:

B0: Average unfished total biomass

Blow: Spawning stock biomass where it takes MGThorizon x MGT to reach Bfrac of BMSY

BMSY: Average total biomass corresponding with MSY

BMSY_B0: Ratio of BMSY to B0

FMSY: Fishing mortality rate corresponding with MSY

FMSY_M: Ratio of FMSY to (adult) M

MGT: Mean generation time

MSY: Maximum sustainable yield

N0: Average unfished numbers

R0: Average unfished recruitment

RefY: Maximum yield obtained in forward projections with a fixed F

SSB0: Average unfished spawning biomass

SSBMSY: Average spawning biomass corresponding with MSY

SSBMSY_SSB: Ratio of SSBMSY to SSB0

UMSY: Exploitation rate corresponding with MSY

VBMSY: Average vulnerable biomass corresponding with MSY

SampPars

All sampled Stock, Fleet, Obs, and Imp parameters

Misc

A list of additional information