Internal SRA MCMC CPP code

Rcpp version of R code

LSRA_MCMC_sim(
  nits,
  pars,
  JumpCV,
  adapt,
  parLB,
  parUB,
  R0ind,
  inflind,
  slpind,
  RDind,
  nyears,
  maxage,
  M,
  Mat_age,
  Wt_age,
  Chist_a,
  Umax,
  h,
  CAA,
  CAAadj,
  sigmaR
)

Arguments

nits
number of iterations

pars
vector of parameters

JumpCV
jump cv vector

adapt
adapt vector

parLB
lower bounds

parUB
upper bounds

R0ind
index for R0

inflind
index for inflection

slpind
index for slope

RDind
index for recruitment deviations

nyears
number of projection years

maxage
maximum age

M
Natural mortality

Mat_age
A vector of maturity at age

Wt_age
A vector of weight at age

Chist_a
A vector of historical catch observations (nyears long) going back to unfished conditions

Umax
A numeric value representing the maximum harvest rate for any age class (rejection of sims where this occurs)

h
steepness of SRR

CAA
A matrix nyears (rows) by nages (columns) of catch at age (age 1 to maxage in length)

CAAadj
internal parameter

sigmaR
A numeric value representing the prior standard deviation of log space recruitment deviations