Class 'MSE'

A Management Strategy Evaluation object that contains information about simulation conditions and performance of data-limited methods

Slots

Name

Name of the MSE object. Single value. Character string

nyears

The number of years for the historical simulation. Single value. Positive integer

proyears

The number of years for the projections - closed loop simulations. Single value. Positive integer

nMPs

Number of management procedures simulation tested. Single value. Positive integer.

MPs

The names of the MPs that were tested. Vector of length nMPs. Character strings.

nsim

Number of simulations. Single value. Positive integer

OM

A table of sampled parameter of the operating model. Table object of nsim rows. Real numbers

A: abundance (biomass) updated in each management update of projection

AC: autocorrelation in recruitment

ageM: age at 50 per cent maturity

B0: unfished total biomass

Blow: SSB where it takes MGThorizon x MGT to reach Bfrac of BMSY (spawning biomass)

BMSY: BMSY in last historical year (total biomass)

BMSY_B0: ratio of BMSY to unfished total biomass in last historical year

Depletion: stock depletion (spawning biomass / unfished spawing biomass) in the final historical year (prior to projection)

dFfinal: gradient in fishing mortality rate over final five years of the historical simulation

DR: the fraction of caught fish that are discarded

Esd: inter-annual variability in fishing mortality rate

Fdisc: fraction of discarded fish that die

FMSY: fishing mortality at MSY in last historical year

FMSY_M: ratio of FMSY to M in last historical year

Frac_area_1: fraction of the unfished biomass in stock 1

hs: steepness of the stock recruitment relationship (the fraction of unfished recruitment at a fifth of unfished stock levels)

K: maximum growth rate (von Bertalanffy K parameter)

Kgrad: mean gradient in maximum growth rate (per cent per time step)

Ksd: interannual variability in maximum growth rate (log normal CV)

L5: length at 5/

L50: length at 50/

L95: length at 95/

LFC: length at first capture

LFR: first length at full retention

LFS: first length at full selection

Linf: asymptotic length

Linfgrad: mean gradient in maximum length (per cent per time step)

Linfsd: interannual variability in maximum length (log normal CV)

LR5: first length at 5/

M: instantaneous natural mortality rate

Mexp: exponent of Lorezen M-weight relationship

Mgrad: mean average percentage gradient in natural mortality rate (percentage per time step)

MGT: mean generation time

Msd: interannual variability in natural mortality rate (lognormal CV)

MSY: maximum sustainable yield in last historical year

N0: equilibrium unfished total numbers

OFLreal: A * FMSY the true simulated Over Fishing Limit in last historical year

Prob_staying: probability of inviduals in area 1 remaining in area 1 over the course of one year

procsd: process error - CV in log-normal recruitment deviations

qcv: interannual variability in future fishing efficiency (catchability) in projected years (input controls only)

qinc: mean percentage increase in fishing efficiency (catchability) in projected years (input controls only)

RefY: reference yield, the highest long-term yield (mean over last five years of projection) obtained from a fixed F strategy. This is a useful reference point for framing performance of MPs because it standardizes for starting point and future productivity.

Rmaxlen: retention of fish at asymptotic length

Size_area_1: size of Area 1 relative to Area 2

SizeLimFrac: fraction of average implementation error of size regulations

SizeLimSD: log-normal coefficient of variation controlling mismatch between a minimum size limit and the real minimum size retained.

Spat_targ: distribution of fishing in relation to spatial biomass: fishing distribution is proportional to B^Spat_targ

SSB0: unfished spawning biomass

SSBMSY: spawning biomass MSY in last historical year

SSBMSY_SSB0: ratio of SSB_MSY/SSB0 in last historical year

t0: theoretical length at age zero (von Bertalanffy t0 parameter)

TACFrac: mean fraction of TAC taken

TACSD: log-normal coefficient of variation in the fraction of Total Allowable Catch (TAC) taken

TAEFrac: mean fraction of TAE taken

TAESD: log-normal coefficient of variation in the fraction of TAE

Vmaxlen: selection of fish at asymptotic length

Obs

A table of sampled parameters of the observation model. Table of nsim rows. Real numbers

Abias: bias in observed current absolute stock biomass

Aerr: error in observed current absolute stock biomass

betas: hyper-stability/hyper-depletion parameter

BMSY_B0bias: bias in ratio of most productive stock size relative to unfished

Brefbias: bias in BMSY stock levels (target or reference biomass levels)

CAA_ESS: the effective sample size of multinomial catch-at-age observation model (number of independent draws)

CAA_nsamp: the number of catch-at-age observations per time step

CAL_ESS: the effective sample size of multinomial catch-at-length observation model (number of independent draws)

CAL_nsamp: the number of catch-at-length observations per time step

Cbias: bias in observed catches

Crefbias: bias in MSY prediction (target or reference catch)

Csd: observation error in observed catches (lognormal CV)

Dbias: bias in observed stock depletion (also applies to depletion Dt for DCAC)

Derr: error in observed stock depletion

FMSY_Mbias: bias in ratio of FMSY to natural mortality rate

hbias: bias in observed steepness of the stock recruitment relationship

Irefbias: bias in abundance index corresponding to BMSY stock levels

Isd: observation error in relative abundance index (lognormal CV)

Kbias: bias in maximum growth rate (von Bertalanffy K parameter)

lenMbias: bias in length at 50 per cent maturity

LFCbias: bias in length at first capture

LFSbias: bias in length at full selection

Linfbias: bias in maximum length (von Bertalanffy Linf parameter)

Mbias: bias in observed natural mortality rate

Recsd: error in observed recruitment

t0bias: bias in theoretical length at age zero (von Bertalanffy t0 parameter)

B_BMSY

Simulated biomass relative to BMSY over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

F_FMSY

Simulated fishing mortality rate relative to FMSY over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

B

Simulated stock biomass over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

SSB

Simulated spawning stock biomass over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

VB

Simulated vulnerable biomass over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

FM

Simulated fishing mortality rate over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

C

Simulated catches (taken) over the projection. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

TAC

Simulated Total Allowable Catch (prescribed) over the projection (this is NA for input controls). An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

SSB_hist

Simulated historical spawning stock biomass. An array with dimensions: nsim, nages, nyears, nareas. Non-negative real numbers

CB_hist

Simulated historical catches in weight. An array with dimensions: nsim, nages, nyears, nareas. Non-negative real numbers

FM_hist

Simulated historical fishing mortality rate. An array with dimensions: nsim, nages, nyears, nareas. Non-negative real numbers

Effort

Simulated relative fishing effort in the projection years. An array with dimensions: nsim, nMPs, proyears. Non-negative real numbers

PAA

Population at age in last projection year. An array with dimensions: nsim, nMPs, nages. Non-negative real numbers

CAA

Catch at age in last projection year. An array with dimensions: nsim, nMPs, nages. Non-negative real numbers

CAL

Catch at length in last projection year. An array with dimensions: nsim, nMPs, nCALbins. Non-negative real numbers

CALbins

Mid-points of the catch-at-length bins. Vector of length nCALbins. Positive real numbers.

Misc

Miscellanenous output such as posterior predictive data

Objects from the Class

Objects can be created by calls of the form new('MSE', Name, nyears, proyears, nMPs, MPs, nsim, OMtable, Obs, B_BMSYa, F_FMSYa, Ba, FMa, Ca, OFLa, Effort, PAA, CAA, CAL, CALbins)