Class 'Obs'

An operating model component that controls the observation model

Slots

Name

The name of the observation model object. Single value. Character string.

Cobs

Log-normal catch observation error expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Non-negative real numbers

Cbiascv

Log-normal coefficient of variation controlling the sampling of bias in catch observations for each simulation. Uniform distribution lower and upper bounds. Non-negative real numbers

CAA_nsamp

Number of catch-at-age observation per time step. Uniform distribution lower and upper bounds. Positive real numbers

CAA_ESS

Effective sample size (independent age draws) of the multinomial catch-at-age observation error model. Uniform distribution lower and upper bounds. Positive integers

CAL_nsamp

Number of catch-at-length observation per time step. Uniform distribution lower and upper bounds. Positive integers

CAL_ESS

Effective sample size (independent length draws) of the multinomial catch-at-length observation error model. Uniform distribution lower and upper bounds. Positive integers

Iobs

Observation error in the relative abundance indices expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Positive real numbers

Ibiascv

Log-normal coefficient of variation controlling error in observations of relative abundance index. Uniform distribution lower and upper bounds. Positive real numbers

Btobs

Log-normal coefficient of variation controlling error in observations of current stock biomass among years. Uniform distribution lower and upper bounds. Positive real numbers

Btbiascv

Uniform-log bounds for sampling persistent bias in current stock biomass. Uniform-log distribution lower and upper bounds. Positive real numbers

beta

A parameter controlling hyperstability/hyperdepletion where values below 1 lead to hyperstability (an index that decreases slower than true abundance) and values above 1 lead to hyperdepletion (an index that decreases more rapidly than true abundance). Uniform distribution lower and upper bounds. Positive real numbers

LenMbiascv

Log-normal coefficient of variation for sampling persistent bias in length at 50 percent maturity. Uniform distribution lower and upper bounds. Positive real numbers

Mbiascv

Log-normal coefficient of variation for sampling persistent bias in observed natural mortality rate. Uniform distribution lower and upper bounds. Positive real numbers

Kbiascv

Log-normal coefficient of variation for sampling persistent bias in observed growth parameter K. Uniform distribution lower and upper bounds. Positive real numbers

t0biascv

Log-normal coefficient of variation for sampling persistent bias in observed t0. Uniform distribution lower and upper bounds. Positive real numbers

Linfbiascv

Log-normal coefficient of variation for sampling persistent bias in observed maximum length. Uniform distribution lower and upper bounds. Positive real numbers

LFCbiascv

Log-normal coefficient of variation for sampling persistent bias in observed length at first capture. Uniform distribution lower and upper bounds. Positive real numbers

LFSbiascv

Log-normal coefficient of variation for sampling persistent bias in length-at-full selection. Uniform distribution lower and upper bounds. Positive real numbers

FMSYbiascv

Log-normal coefficient of variation for sampling persistent bias in FMSY. Uniform distribution lower and upper bounds. Positive real numbers

FMSY_Mbiascv

Log-normal coefficient of variation for sampling persistent bias in FMSY/M. Uniform distribution lower and upper bounds. Positive real numbers

BMSY_B0biascv

Log-normal coefficient of variation for sampling persistent bias in BMSY relative to unfished. Uniform distribution lower and upper bounds. Positive real numbers

Irefbiascv

Log-normal coefficient of variation for sampling persistent bias in relative abundance index at BMSY. Uniform distribution lower and upper bounds. Positive real numbers

Brefbiascv

Log-normal coefficient of variation for sampling persistent bias in BMSY. Uniform distribution lower and upper bounds. Positive real numbers

Crefbiascv

Log-normal coefficient of variation for sampling persistent bias in MSY. Uniform distribution lower and upper bounds. Positive real numbers

Dbiascv

Log-normal coefficient of variation for sampling persistent bias in stock depletion. Uniform distribution lower and upper bounds. Positive real numbers

Dobs

Log-normal coefficient of variation controlling error in observations of stock depletion among years. Uniform distribution lower and upper bounds. Positive real numbers

hbiascv

Log-normal coefficient of variation for sampling persistent bias in steepness. Uniform distribution lower and upper bounds. Positive real numbers

Recbiascv

Log-normal coefficient of variation for sampling persistent bias in recent recruitment strength. Uniform distribution lower and upper bounds. Positive real numbers

Note

Its questionable whether the hyperstability/hyperdepletion should be categorised as an observation model characteristic as it is most often driven by fleet dynamics (and therefore should be in the fleet object). Oh well its here and you might want to make it hyperstable beta < 1 or hyperdeplete beta > 1, only.

Objects from the Class

Objects can be created by calls of the form new('Obs')

Examples

showClass('Obs')

#> Class "Obs" [package "DLMtool"]
#> 
#> Slots:
#>                                                                             
#> Name:           Name          Cobs       Cbiascv     CAA_nsamp       CAA_ESS
#> Class:     character       numeric       numeric       numeric       numeric
#>                                                                             
#> Name:      CAL_nsamp       CAL_ESS          Iobs       Ibiascv         Btobs
#> Class:       numeric       numeric       numeric       numeric       numeric
#>                                                                             
#> Name:       Btbiascv          beta    LenMbiascv       Mbiascv       Kbiascv
#> Class:       numeric       numeric       numeric       numeric       numeric
#>                                                                             
#> Name:       t0biascv    Linfbiascv     LFCbiascv     LFSbiascv    FMSYbiascv
#> Class:       numeric       numeric       numeric       numeric       numeric
#>                                                                             
#> Name:   FMSY_Mbiascv BMSY_B0biascv    Irefbiascv    Brefbiascv    Crefbiascv
#> Class:       numeric       numeric       numeric       numeric       numeric
#>                                                               
#> Name:        Dbiascv          Dobs       hbiascv     Recbiascv
#> Class:       numeric       numeric       numeric       numeric
#> 
#> Known Subclasses: "OM"