Performance Metrics

Performance metrics (PMs) for your management strategy evaluation.

P10(MSEobj = NULL, Ref = 0.1, Yrs = NULL)

P50(MSEobj = NULL, Ref = 0.5, Yrs = NULL)

P100(MSEobj = NULL, Ref = 1, Yrs = NULL)

POF(MSEobj = NULL, Ref = 1, Yrs = NULL)

LTY(MSEobj = NULL, Ref = 0.5, Yrs = c(max(MSEobj@proyears - 9, 1),
  MSEobj@proyears))

STY(MSEobj = NULL, Ref = 0.5, Yrs = c(1, 10))

Yield(MSEobj = NULL, Ref = 1, Yrs = NULL)

AAVY(MSEobj = NULL, Ref = 0.2, Yrs = NULL)

Arguments

MSEobj
An object of class MSE

Ref
Reference point for calculating the performance metric. See details.

Yrs
Numeric vector of length 2 with year indices to summarize performance. If NULL, the performance is summarized over all projection years.

Value

An object of class PMobj

Details

Performance Metric definitions:

P10	Probability B > 0.1 BMSY

P50	Probability B > 0.5 BMSY

P100	Probability B > BMSY

POF	Probability F < FMSY

LTY	Probability Long-Term Yield > 0.5 Relative Yield

STY	Probability Short-Term Yield > 0.5 Relative Yield

AAVY	Probability AAVY < 0.2 (Average Annual Variability in Yield)

Yield	Average Yield (relative to Reference Yield)

Argument Ref provides the ratio relative to the reference point for calculating the performance metric. For biomass-based PMs (P10, P50, P100), this is the fraction of BMSY. For POF, the fraction of FMSY. For Yield (and LTY/STY), the fraction of the Reference Yield.

Long-Term Yield is the Yield in the last ten years of the projection period in the MSE. Short-Term Yield is that in the first 10 years of the projection period

Examples

# NOT RUN {
P10(myMSE)
P50(myMSE)
P100(myMSE)
POF(myMSE)
LTY(myMSE)
STY(myMSE)
AAVY(myMSE)
Yield(myMSE)
# }