What management procedures need what data

A function that finds all the MPs and searches the function text for slots in the Data object

Required(funcs = NA, noCV = FALSE)

Arguments

funcs
A character vector of management procedures

noCV
Logical. Should the CV slots be left out?

Value

A matrix of MPs and their required data in terms of Data slotnames

See also

Examples

Required(c("DCAC", "AvC"))

#>    MP    
#> 19 "DCAC"
#> 1  "AvC" 
#>    Data                                                                                    
#> 19 "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"
#> 1  "Cat, LHYear, t, Year"                                                                  

Required() # For all MPs

#>        MP            
#>   [1,] "AvC"         
#>   [2,] "AvC_MLL"     
#>   [3,] "BK"          
#>   [4,] "BK_CC"       
#>   [5,] "BK_ML"       
#>   [6,] "CC1"         
#>   [7,] "CC2"         
#>   [8,] "CC3"         
#>   [9,] "CC4"         
#>  [10,] "CC5"         
#>  [11,] "CompSRA"     
#>  [12,] "CompSRA4010" 
#>  [13,] "curE"        
#>  [14,] "curE75"      
#>  [15,] "DAAC"        
#>  [16,] "DBSRA"       
#>  [17,] "DBSRA_40"    
#>  [18,] "DBSRA4010"   
#>  [19,] "DCAC"        
#>  [20,] "DCAC_40"     
#>  [21,] "DCAC_ML"     
#>  [22,] "DCAC4010"    
#>  [23,] "DCACs"       
#>  [24,] "DD"          
#>  [25,] "DD4010"      
#>  [26,] "DDe"         
#>  [27,] "DDe75"       
#>  [28,] "DDes"        
#>  [29,] "DepF"        
#>  [30,] "DTe40"       
#>  [31,] "DTe50"       
#>  [32,] "DynF"        
#>  [33,] "EtargetLopt" 
#>  [34,] "Fadapt"      
#>  [35,] "Fdem"        
#>  [36,] "Fdem_CC"     
#>  [37,] "Fdem_ML"     
#>  [38,] "FMSYref"     
#>  [39,] "FMSYref50"   
#>  [40,] "FMSYref75"   
#>  [41,] "Fratio"      
#>  [42,] "Fratio_CC"   
#>  [43,] "Fratio_ML"   
#>  [44,] "Fratio4010"  
#>  [45,] "GB_CC"       
#>  [46,] "GB_slope"    
#>  [47,] "GB_target"   
#>  [48,] "Gcontrol"    
#>  [49,] "HDAAC"       
#>  [50,] "ICI"         
#>  [51,] "ICI2"        
#>  [52,] "Iratio"      
#>  [53,] "Islope1"     
#>  [54,] "Islope4"     
#>  [55,] "IT10"        
#>  [56,] "IT5"         
#>  [57,] "Itarget1"    
#>  [58,] "Itarget1_MPA"
#>  [59,] "Itarget4"    
#>  [60,] "ItargetE1"   
#>  [61,] "ItargetE4"   
#>  [62,] "ITe10"       
#>  [63,] "ITe5"        
#>  [64,] "ITM"         
#>  [65,] "L95target"   
#>  [66,] "Lratio_BHI"  
#>  [67,] "Lratio_BHI2" 
#>  [68,] "LstepCC1"    
#>  [69,] "LstepCC4"    
#>  [70,] "LstepCE1"    
#>  [71,] "LstepCE2"    
#>  [72,] "Ltarget1"    
#>  [73,] "Ltarget4"    
#>  [74,] "LtargetE1"   
#>  [75,] "LtargetE4"   
#>  [76,] "matlenlim"   
#>  [77,] "matlenlim2"  
#>  [78,] "MCD"         
#>  [79,] "MCD4010"     
#>  [80,] "minlenLopt1" 
#>  [81,] "MRnoreal"    
#>  [82,] "MRreal"      
#>  [83,] "NFref"       
#>  [84,] "Rcontrol"    
#>  [85,] "Rcontrol2"   
#>  [86,] "SBT1"        
#>  [87,] "SBT2"        
#>  [88,] "slotlim"     
#>  [89,] "SPmod"       
#>  [90,] "SPMSY"       
#>  [91,] "SPslope"     
#>  [92,] "SPSRA"       
#>  [93,] "SPSRA_ML"    
#>  [94,] "YPR"         
#>  [95,] "YPR_CC"      
#>  [96,] "YPR_ML"      
#>  [97,] "Islope2"     
#>  [98,] "Itarget2"    
#>  [99,] "Itarget3"    
#> [100,] "LstepCC2"    
#> [101,] "LstepCC3"    
#> [102,] "Ltarget2"    
#> [103,] "Ltarget3"    
#>        Data                                                                                                                                                                                                                                                                                                                                                                                                                   
#>   [1,] "Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                                 
#>   [2,] "Cat, L50, t"                                                                                                                                                                                                                                                                                                                                                                                                          
#>   [3,] "Abun, CV_Abun, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                                                                   
#>   [4,] "CAA, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, LFC, Mort, t, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                                              
#>   [5,] "CAL, CAL_bins, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, LFC, LFS, Mort, t, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                               
#>   [6,] "Cat, CV_Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                         
#>   [7,] "Cat, CV_Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                         
#>   [8,] "Cat, CV_Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                         
#>   [9,] "Cat, CV_Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                         
#>  [10,] "Cat, CV_Cat, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                         
#>  [11,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                  
#>  [12,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                  
#>  [13,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [14,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [15,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                               
#>  [16,] "BMSY_B0, Cat, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, t, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                              
#>  [17,] "BMSY_B0, Cat, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, t, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                              
#>  [18,] "BMSY_B0, Cat, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, t, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                              
#>  [19,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                               
#>  [20,] "AvC, BMSY_B0, CV_BMSY_B0, CV_FMSY_M, CV_Mort, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                                          
#>  [21,] "AvC, BMSY_B0, CAL, CAL_bins, Cat, CV_BMSY_B0, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, LHYear, Mort, t, vbK, vbLinf, Year"                                                                                                                                                                                                                                                                                 
#>  [22,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                               
#>  [23,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                               
#>  [24,] "Cat, Ind, L50, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                                                    
#>  [25,] "Cat, Ind, L50, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                                                    
#>  [26,] "Cat, Ind, L50, LHYear, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                                            
#>  [27,] "Cat, Ind, L50, LHYear, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                                            
#>  [28,] "Cat, Ind, L50, LHYear, MaxAge, Mort, MPeff, t, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                                     
#>  [29,] "Abun, BMSY_B0, CV_Abun, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort, t, Year"                                                                                                                                                                                                                                                                                                                                       
#>  [30,] "Dep, MPeff"                                                                                                                                                                                                                                                                                                                                                                                                           
#>  [31,] "Dep, MPeff"                                                                                                                                                                                                                                                                                                                                                                                                           
#>  [32,] "Abun, Cat, CV_FMSY_M, CV_Mort, FMSY_M, Ind, Mort, t, Year"                                                                                                                                                                                                                                                                                                                                                            
#>  [33,] "L50, ML, Mort, MPeff, t, vbK, vbLinf, wlb, Year"                                                                                                                                                                                                                                                                                                                                                                      
#>  [34,] "Abun, Cat, FMSY_M, Ind, Mort, t, Year"                                                                                                                                                                                                                                                                                                                                                                                
#>  [35,] "Abun, CV_Abun, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                                    
#>  [36,] "CAA, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                                 
#>  [37,] "CAL, CAL_bins, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                  
#>  [38,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [39,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [40,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [41,] "Abun, CV_Abun, CV_FMSY_M, CV_Mort, FMSY_M, Mort, t"                                                                                                                                                                                                                                                                                                                                                                   
#>  [42,] "CAA, Cat, CV_Cat, CV_FMSY_M, CV_Mort, FMSY_M, Mort, t"                                                                                                                                                                                                                                                                                                                                                                
#>  [43,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, t, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                 
#>  [44,] "Abun, CV_Abun, CV_Dep, CV_Dt, CV_FMSY_M, CV_Mort, Dep, Dt, FMSY_M, Mort, t"                                                                                                                                                                                                                                                                                                                                           
#>  [45,] "Cat, Cref, CV_Cref, t"                                                                                                                                                                                                                                                                                                                                                                                                
#>  [46,] "Cat, CV_Cat, Ind, t, Year"                                                                                                                                                                                                                                                                                                                                                                                            
#>  [47,] "Cat, Cref, CV_Cref, CV_Iref, Ind, Iref, t"                                                                                                                                                                                                                                                                                                                                                                            
#>  [48,] "Abun, Cat, Ind, t, Year"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [49,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                                               
#>  [50,] "Cat, CV_Cat, CV_Ind, Ind, Year"                                                                                                                                                                                                                                                                                                                                                                                       
#>  [51,] "Cat, CV_Cat, CV_Ind, Ind, t"                                                                                                                                                                                                                                                                                                                                                                                          
#>  [52,] "Cat, CV_Cat, CV_Ind, Ind, t, Year"                                                                                                                                                                                                                                                                                                                                                                                    
#>  [53,] "Cat, CV_Cat, Ind, LHYear, MPrec, t, Year"                                                                                                                                                                                                                                                                                                                                                                             
#>  [54,] "Cat, CV_Cat, Ind, LHYear, MPrec, t, Year"                                                                                                                                                                                                                                                                                                                                                                             
#>  [55,] "CV_Ind, Ind, Iref, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                                                       
#>  [56,] "CV_Ind, Ind, Iref, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                                                       
#>  [57,] "Cat, CV_Cat, Ind, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                    
#>  [58,] "Cat, CV_Cat, Ind, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                    
#>  [59,] "Cat, CV_Cat, Ind, LHYear, t, Year"                                                                                                                                                                                                                                                                                                                                                                                    
#>  [60,] "Ind, LHYear, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                             
#>  [61,] "Ind, LHYear, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                             
#>  [62,] "CV_Ind, Ind, Iref, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                       
#>  [63,] "CV_Ind, Ind, Iref, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                       
#>  [64,] "Cat, CV_Ind, Ind, Iref, Mort, MPrec, t, Year"                                                                                                                                                                                                                                                                                                                                                                         
#>  [65,] "Cat, CV_Cat, L50, L95, LHYear, ML, t, Year"                                                                                                                                                                                                                                                                                                                                                                           
#>  [66,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_vbLinf, LFS, t, vbLinf"                                                                                                                                                                                                                                                                                                                                                        
#>  [67,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, t, vbK, vbLinf"                                                                                                                                                                                                                                                                                                         
#>  [68,] "Cat, CV_Cat, LHYear, ML, MPrec, t, Year"                                                                                                                                                                                                                                                                                                                                                                              
#>  [69,] "Cat, CV_Cat, LHYear, ML, MPrec, t, Year"                                                                                                                                                                                                                                                                                                                                                                              
#>  [70,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [71,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [72,] "Cat, CV_Cat, LHYear, ML, t, Year"                                                                                                                                                                                                                                                                                                                                                                                     
#>  [73,] "Cat, CV_Cat, LHYear, ML, t, Year"                                                                                                                                                                                                                                                                                                                                                                                     
#>  [74,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [75,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [76,] "L50"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [77,] "L50"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [78,] "Cat, CV_Cat, CV_Dep, Dep, t"                                                                                                                                                                                                                                                                                                                                                                                          
#>  [79,] "Cat, CV_Cat, CV_Dep, Dep, t"                                                                                                                                                                                                                                                                                                                                                                                          
#>  [80,] "Mort, MPeff, t, vbK, vbLinf, wlb"                                                                                                                                                                                                                                                                                                                                                                                     
#>  [81,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [82,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [83,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [84,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, Ind, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, Year"                                                                                                                                                                                                                                                                                          
#>  [85,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, Ind, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0, Year"                                                                                                                                                                                                                                                                                          
#>  [86,] "Cat, CV_Cat, Ind, t, Year"                                                                                                                                                                                                                                                                                                                                                                                            
#>  [87,] "Cat, Cref, CV_Cref, Rec, t"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [88,] "L50, vbLinf"                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [89,] "Abun, Cat, CV_Abun, CV_Cat, CV_Ind, Ind, t"                                                                                                                                                                                                                                                                                                                                                                           
#>  [90,] "Cat, L50, MaxAge, t, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                                                               
#>  [91,] "Abun, Cat, CV_Cat, Ind, t, Year"                                                                                                                                                                                                                                                                                                                                                                                      
#>  [92,] "Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, MaxAge, Mort, steep, t, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                   
#>  [93,] "Abun, AvC, BMSY_B0, Bref, CAL, CAL_bins, Cat, Cref, CV_AvC, CV_BMSY_B0, CV_Bref, CV_Cat, CV_Cref, CV_Dep, CV_Dt, CV_FMSY_M, CV_Ind, CV_Iref, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_Rec, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, Dep, Dt, FMSY_M, Ind, Iref, L50, L95, Lbar, Lc, LenCV, LFC, LFS, LHYear, MaxAge, ML, Mort, MPeff, MPrec, Rec, sigmaL, SpAbun, steep, t, vbK, vbLinf, vbt0, wla, wlb, Year"
#>  [94,] "Abun, CV_Abun, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                                        
#>  [95,] "CAA, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                                     
#>  [96,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, t, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                                           
#>  [97,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [98,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [99,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [100,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [101,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [102,] ""                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [103,] ""