Class 'Stock'

An operating model component that specifies the parameters of the population dynamics model

Slots

Name

The name of the Stock object. Single value. Character string

Common_Name

Common name of the species. Character string

Species

Scientific name of the species. Genus and species name. Character string

maxage

The maximum age of individuals that is simulated (there is no 'plus group'). Single value. Positive integer

R0

The magnitude of unfished recruitment. Single value. Positive real number

M

Natural mortality rate. Uniform distribution lower and upper bounds. Positive real number

M2

(Optional) Natural mortality rate at age. Vector of length 'maxage'. Positive real number

Mexp

Exponent of the Lorenzen function assuming an inverse relationship between M and weight. Uniform distribution lower and upper bounds. Real numbers <= 0.

Msd

Inter-annual variability in natural mortality rate expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Non-negative real numbers

Mgrad

No longer used. Previously mean temporal trend in natural mortality rate, expressed as a percentage change in M per year.

h

Steepness of the stock recruit relationship. Uniform distribution lower and upper bounds. Values from 1/5 to 1

SRrel

Type of stock-recruit relationship. Single value, switch (1) Beverton-Holt (2) Ricker. Integer

Perr

Process error, the CV of lognormal recruitment deviations. Uniform distribution lower and upper bounds. Non-negative real numbers

AC

Autocorrelation in recruitment deviations rec(t)=AC*rec(t-1)+(1-AC)*sigma(t). Uniform distribution lower and upper bounds. Non-negative real numbers

Period

(Optional) Period for cyclical recruitment pattern in years. Uniform distribution lower and upper bounds. Non-negative real numbers

Amplitude

(Optional) Amplitude in deviation from long-term average recruitment during recruitment cycle (eg a range from 0 to 1 means recruitment decreases or increases by up to 100% each cycle). Uniform distribution lower and upper bounds. 0 < Amplitude < 1

Linf

Maximum length. Uniform distribution lower and upper bounds. Positive real numbers

K

von Bertalanffy growth parameter k. Uniform distribution lower and upper bounds. Positive real numbers

t0

von Bertalanffy theoretical age at length zero. Uniform distribution lower and upper bounds. Non-positive real numbers

LenCV

Coefficient of variation of length-at-age (assumed constant for all age classes). Uniform distribution lower and upper bounds. Positive real numbers

Ksd

Inter-annual variability in growth parameter k expressed as coefficient of variation. Uniform distribution lower and upper bounds. Non-negative real numbers

Kgrad

No longer used. Previously mean temporal trend in growth parameter k, expressed as a percentage change in k per year.

Linfsd

Inter-annual variability in maximum length expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Non-negative real numbers

Linfgrad

No longer used. Previously mean temporal trend in maximum length, expressed as a percentage change in Linf per year.

L50

Length at 50 percent maturity. Uniform distribution lower and upper bounds. Positive real numbers

L50_95

Length increment from 50 percent to 95 percent maturity. Uniform distribution lower and upper bounds. Positive real numbers

D

Current level of stock depletion SSB(current)/SSB(unfished). Uniform distribution lower and upper bounds. Fraction

a

Length-weight parameter alpha. Single value. Positive real number

b

Length-weight parameter beta. Single value. Positive real number

Size_area_1

The size of area 1 relative to area 2. Uniform distribution lower and upper bounds. Positive real numbers

Frac_area_1

The fraction of the unfished biomass in stock 1. Uniform distribution lower and upper bounds. Positive real numbers

Prob_staying

The probability of inviduals in area 1 remaining in area 1 over the course of one year. Uniform distribution lower and upper bounds. Positive fraction.

Fdisc

Fraction of discarded fish that die. Uniform distribution lower and upper bounds. Non-negative real numbers

Source

A reference to a website or article from which parameters were taken to define the stock object. Single value. Character string.

Objects from the Class

Objects can be created by calls of the form new('Stock')

Examples

showClass('Stock')

#> Class "Stock" [package "DLMtool"]
#> 
#> Slots:
#>                                                                        
#> Name:          Name  Common_Name      Species       maxage           R0
#> Class:    character    character    character      numeric      numeric
#>                                                                        
#> Name:             M           M2         Mexp          Msd        Mgrad
#> Class:      numeric      numeric      numeric      numeric      numeric
#>                                                                        
#> Name:             h        SRrel         Perr           AC       Period
#> Class:      numeric      numeric      numeric      numeric      numeric
#>                                                                        
#> Name:     Amplitude         Linf            K           t0        LenCV
#> Class:      numeric      numeric      numeric      numeric      numeric
#>                                                                        
#> Name:           Ksd        Kgrad       Linfsd     Linfgrad          L50
#> Class:      numeric      numeric      numeric      numeric      numeric
#>                                                                        
#> Name:        L50_95            D            a            b  Size_area_1
#> Class:      numeric      numeric      numeric      numeric      numeric
#>                                                           
#> Name:   Frac_area_1 Prob_staying        Fdisc       Source
#> Class:      numeric      numeric      numeric    character
#> 
#> Known Subclasses: "OM"