Optimize yield for a single simulation

Optimize yield for a single simulation

optMSY(logFa, Asize_c, nareas, maxage, Ncurr, pyears, M_age, MatAge, WtAge,
  Vuln, Retc, Prec, movc, SRrelc, Effind, Spat_targc, hc, R0c, SSBpRc, aRc, bRc,
  Qc, MPA, maxF, useCPP = TRUE, SSB0c)

Arguments

logFa
log apical fishing mortality

Asize_c
A vector of length areas with relative size of areas

nareas
Number of area

maxage
Maximum age

Ncurr
Current N-at-age

pyears
Number of projection years

M_age
M-at-age

MatAge
Maturity-at-age

WtAge
Weight-at-age

Vuln
Vulnerablity-at-age

Retc
Retention-at-age

Prec
Recruitment error

movc
Movement matrix

SRrelc
SR Relationship

Effind
Historical effort

Spat_targc
Spatial targeting

hc
Steepness

R0c
Unfished recruitment by area

SSBpRc
Unfished spawning stock per recruit by area

aRc
Ricker aR

bRc
Ricker bR

Qc
Catchability

MPA
A matrix of spatial closures by year

maxF
A numeric value specifying the maximum fishing mortality for any single age class

useCPP
logical - use the CPP code? For testing purposes only

SSB0c
SSB0