Simulate population dynamics for historical years

Simulate population dynamics for historical years

simYears(x, nareas, maxage, N, pyears, M_ageArray, Asize, Mat_age, Wt_age, V,
  retA, Perr, mov, SRrel, Find, Spat_targ, hs, R0a, SSBpR, aR, bR, qs, MPA,
  maxF, useCPP = TRUE, SSB0)

Arguments

x
Integer, the simulation number

nareas
The number of spatial areas

maxage
The maximum age

N
Array of the numbers-at-age in population. Dimensions are nsim, maxage, nyears, nareas. Only values from the first year (i.e N[,,1,]) are used, which is the current N-at-age.

pyears
The number of years to project forward. Equal to 'nyears' for optimizing for q.

M_ageArray
An array (dimensions nsim, maxage, nyears+proyears) with the natural mortality-at-age and year

Asize
A matrix (dimensions nsim, nareas) of size of areas

Mat_age
A matrix (dimensions nsim, maxage) with the proportion mature for each age-class

Wt_age
An array (dimensions nsim, maxage, nyears+proyears) with the weight-at-age and year

V
An array (dimensions nsim, maxage, nyears+proyears) with the vulnerability-at-age and year

retA
An array (dimensions nsim, maxage, nyears+proyears) with the probability retained-at-age and year

Perr
A matrix (dimensions nsim, nyears+proyears) with the recruitment deviations

mov
An array (dimensions nsim, nareas, nareas) with the movement matrix

SRrel
A numeric vector nsim long specifying the recruitment curve to use

Find
A matrix (dimensions nsim, nyears) with the historical fishing effort

Spat_targ
A numeric vector nsim long with the spatial targeting

hs
A numeric vector nsim long with the steepness values for each simulation

R0a
A matrix (dimensions nsim, nareas) with the unfished recruitment by area

SSBpR
A matrix (dimensions nsim, nareas) with the unfished spawning-per-recruit by area

aR
A numeric vector nsim long with the Ricker SRR a values

bR
A numeric vector nsim long with the Ricker SRR b values

qs
A numeric vector nsim long with catchability coefficients

MPA
A matrix of spatial closures by year

maxF
A numeric value specifying the maximum fishing mortality for any single age class

useCPP
logical - use the CPP code? For testing purposes only

SSB0
SSB0