DLMtool: Data-Limited Methods Toolkit

A collection of data-limited management procedures that can be evaluated with management strategy evaluation with the 'MSEtool' package, or applied to fishery data to provide management recommendations.

Author

Maintainer: Adrian Hordyk adrian@bluematterscience.com

Authors:

Tom Carruthers tom@bluematterscience.com

Quang Huynh quang@bluematterscience.com

Other contributors:

M. Bryan [contributor]

HF. Geremont [contributor]

C. Grandin [contributor]

W. Harford [contributor]

Q. Huynh [contributor]

C. Walters [contributor]