Class 'Data'

An object for storing data for analysis using data-limited methods

Slots

Name

The name of the Data object. Single value. Character string

Year

Years that corresponding to catch and relative abundance data. Vector nyears long. Positive integer

Cat

Total annual catches. Matrix of nsim rows and nyears columns. Non-negative real numbers

Ind

Relative abundance index. Matrix of nsim rows and nyears columns. Non-negative real numbers

Rec

Recent recruitment strength. Vector of length nsim. Positive real numbers

t

The number of years corresponding to AvC and Dt. Single value. Positive integer

AvC

Average catch over time t. Vector nsim long. Positive real numbers

Dt

Depletion over time t SSB(now)/SSB(now-t+1). Vector nsim long. Fraction

Mort

Natural mortality rate. Vector nsim long. Positive real numbers

FMSY_M

An assumed ratio of FMSY to M. Vector nsim long. Positive real numbers

BMSY_B0

The most productive stock size relative to unfished. Vector nsim long. Fraction

L50

Length at 50 percent maturity. Vector nsim long. Positive real numbers

L95

Length at 95 percent maturity. Vector nsim long. Positive real numbers

ML

Mean length time series. Matrix of nsim rows and nyears columns. Non-negative real numbers

Lbar

Mean length of catches over Lc. Matrix of nsim rows and nyears columns. Positive real numbers

Lc

Modal length of catches. Matrix of nsim rows and nyears columns. Positive real numbers

LFC

Length at first capture. Matrix of nsim rows and nyears columns. Positive real numbers

LFS

Shortest length at full selection. Matrix of nsim rows and nyears columns. Positive real numbers

CAA

Catch at Age data. Array of dimensions nsim x nyears x MaxAge. Non-negative integers

Dep

Stock depletion SSB(current)/SSB(unfished). Vector nsim long. Fraction.

Abun

An estimate of absolute current vulnerable abundance. Vector nsim long. Positive real numbers

SpAbun

An estimate of absolute current spawning stock abundance. Vector nsim long. Positive real numbers

vbK

The von Bertalanffy growth coefficient K. Vector nsim long. Positive real numbers

vbLinf

Maximum length. Vector nsim long. Positive real numbers

vbt0

Theoretical age at length zero. Vector nsim long. Non-positive real numbers

LenCV

Coefficient of variation of length-at-age (assumed constant for all age classes). Vector nsim long. Positive real numbers

wla

Weight-Length parameter alpha. Vector nsim long. Positive real numbers

wlb

Weight-Length parameter beta. Vector nsim long. Positive real numbers

steep

Steepness of stock-recruitment relationship. Vector nsim long. Value in the range of one-fifth to 1

CV_Cat

Coefficient of variation in annual catches. Vector nsim long. Positive real numbers

CV_Dt

Coefficient of variation in depletion over time t. Vector nsim long. Positive real numbers

CV_AvC

Coefficient of variation in average catches over time t. Vector nsim long. Positive real numbers

CV_Ind

Coefficient of variation in the relative abundance index. Vector nsim long. Positive real numbers

CV_Mort

Coefficient of variation in natural mortality rate. Vector nsim long. Positive real numbers

CV_FMSY_M

Coefficient of variation in the ratio in FMSY/M. Vector nsim long. Positive real numbers

CV_BMSY_B0

Coefficient of variation in the position of the most productive stock size relative to unfished. Vector nsim long. Positive real numbers

CV_Dep

Coefficient of variation in current stock depletion. Vector nsim long. Positive real numbers

CV_Abun

Coefficient of variation in estimate of absolute current stock size. Vector nsim long. Positive real numbers

CV_vbK

Coefficient of variation in the von Bertalanffy K parameter. Vector nsim long. Positive real numbers

CV_vbLinf

Coefficient of variation in maximum length. Vector nsim long. Positive real numbers

CV_vbt0

Coefficient of variation in age at length zero. Vector nsim long. Positive real numbers

CV_L50

Coefficient of variation in length at 50 per cent maturity. Vector nsim long. Positive real numbers

CV_LFC

Coefficient of variation in length at first capture. Vector nsim long. Positive real numbers

CV_LFS

Coefficient of variation in length at full selection. Vector nsim long. Positive real numbers

CV_wla

Coefficient of variation in weight-length parameter a. Vector nsim long. Positive real numbers

CV_wlb

Coefficient of variation in weight-length parameter b. Vector nsim long. Positive real numbers

CV_steep

Coefficient of variation in steepness. Vector nsim long. Positive real numbers

sigmaL

Assumed observaton error of the length composition data. Vector nsim long. Positive real numbers

MaxAge

Maximum age. Vector nsim long. Positive integer

CAL_bins

The values delimiting the length bins for the catch-at-length data. Vector. Non-negative real numbers

CAL

Catch-at-length data. An array with dimensions nsim x nyears x length(CAL_bins). Non-negative integers

TAC

The calculated catch limits (function TAC). An array with dimensions PosMPs x replicate TAC samples x nsim. Positive real numbers

Sense

The results of the sensitivity analysis (function Sense). An array with dimensions PosMPs x sensitivity increments. Positive real numbers

Units

Units of the catch/absolute abundance estimates. Single value. Character string

Ref

A reference management level (eg a catch limit). Single value. Positive real number

Ref_type

Type of reference management level (eg 2009 catch limit). Single value. Character string

Log

A record of events. Single value. Character string

params

A place to store estimated parameters. An object. R list

PosMPs

The methods that can be applied to these data. Vector. Character strings

MPs

The methods that were applied to these data. Vector. Character strings

OM

A table of operating model conditions. R table object of nsim rows. Real numbers

Obs

A table of observation model conditions. R table object of nsim rows. Real numbers

Cref

Reference or target catch level (eg MSY). Vector of length nsim. Positive real numbers

Iref

Reference or target relative abundance index level (eg BMSY / B0). Vector of length nsim. Positive real numbers

Bref

Reference or target biomass level (eg BMSY). Vector of length nsim. Positive real numbers

CV_Cref

Log-normal CV for reference or target catch level. Vector of length nsim. Positive real numbers

CV_Iref

Log-normalCV for reference or target relative abundance index level. Vector of length nsim. Positive real numbers

CV_Bref

Log-normal CV for reference or target biomass level. Vector of length nsim. Positive real numbers

CV_Rec

Log-normal CV for recent recruitment strength. Vector of length nsim. Positive real numbers

MPrec

The previous recommendation of a management procedure. Vector of length nsim. Positive real numbers

MPeff

The current level of effort. Vector of length nsim. Positive real numbers

LHYear

The last historical year of the simulation (before projection). Single value. Positive integer

nareas

Number of fishing areas. Vector of length nsim. Non-negative integer

Misc

Other information for MPs. An object. R list

Objects from the Class

Objects can be created by calls of the form new('Data', stock)

Examples

newdata<-new('Data')

#> [1] "Couldn't find specified csv file, blank DLM object created"