Class 'Fleet'

The component of the operating model that controls fishing dynamics

Slots

Name

Name of the Fleet object. Single value. Character string.

nyears

The number of years for the historical 'spool-up' simulation. Single value. Positive integer

Spat_targ

Distribution of fishing in relation to spatial biomass: fishing distribution is proportional to B^Spat_targ. Uniform distribution lower and upper bounds. Real numbers

EffYears

Years representing join-points (vertices) of time-varying effort. Vector. Non-negative real numbers

EffLower

Lower bound on relative effort corresponding to EffYears. Vector. Non-negative real numbers

EffUpper

Upper bound on relative effort corresponding to EffYears. Vector. Non-negative real numbers

Esd

Additional inter-annual variability in fishing mortality rate. Uniform distribution lower and upper bounds. Non-negative real numbers

qinc

Average percentage change in fishing efficiency (applicable only to forward projection and input controls). Uniform distribution lower and upper bounds. Non-negative real numbers

qcv

Inter-annual variability in fishing efficiency (applicable only to forward projection and input controls). Uniform distribution lower and upper bounds. Non-negative real numbers

L5

Shortest length corresponding to 5 percent vulnerability. Uniform distribution lower and upper bounds. Positive real numbers

LFS

Shortest length that is fully vulnerable to fishing. Uniform distribution lower and upper bounds. Positive real numbers

Vmaxlen

The vulnerability of fish at Stock@Linf. Uniform distribution lower and upper bounds. Fraction

isRel

Selectivity parameters in units of size-of-maturity (or absolute eg cm). Single value. Boolean.

LR5

Shortest length corresponding ot 5 percent retention. Uniform distribution lower and upper bounds. Non-negative real numbers

LFR

Shortest length that is fully retained. Uniform distribution lower and upper bounds. Non-negative real numbers

Rmaxlen

The retention of fish at Stock@Linf. Uniform distribution lower and upper bounds. Non-negative real numbers

DR

Discard rate - the fraction of caught fish that are discarded. Uniform distribution lower and upper bounds. Fraction

SelYears

(Optional) Years representing join-points (vertices) at which historical selectivity pattern changes. Vector. Positive real numbers

AbsSelYears

(Optional) Calendar years corresponding with SelYears (eg 1951, rather than 1), used for plotting only. Vector (of same length as SelYears). Positive real numbers

L5Lower

(Optional) Lower bound of L5 (use ChooseSelect function to set these). Vector. Non-negative real numbers

L5Upper

(Optional) Upper bound of L5 (use ChooseSelect function to set these). Vector. Non-negative real numbers

LFSLower

(Optional) Lower bound of LFS (use ChooseSelect function to set these). Vector. Non-negative real numbers

LFSUpper

(Optional) Upper bound of LFS (use ChooseSelect function to set these). Vector. Non-negative real numbers

VmaxLower

(Optional) Lower bound of Vmaxlen (use ChooseSelect function to set these). Vector. Fraction

VmaxUpper

(Optional) Upper bound of Vmaxlen (use ChooseSelect function to set these). Vector. Fraction

CurrentYr

The current calendar year (final year) of the historical simulations (eg 2011). Single value. Positive integer.

MPA

(Optional) Matrix specifying spatial closures for historical years.

Creating Object

Objects can be created by calls of the form new('Fleet')

MPA slot

Each row should contain year index (e.g 10 for 10th historical year) followed by fraction of area closed to fishing for each area. i.e. each row represents a change and the number of columns is nareas + 1. The spatial closures are assumed to remain in place for the future projections unless changed by a MP. Default (if left blank) is all areas are open to fishing in historical period.

Examples

showClass('Fleet')

#> Class "Fleet" [package "DLMtool"]
#> 
#> Slots:
#>                                                                               
#> Name:         Name      nyears   Spat_targ    EffYears    EffLower    EffUpper
#> Class:   character     numeric     numeric     numeric     numeric     numeric
#>                                                                               
#> Name:          Esd        qinc         qcv          L5         LFS     Vmaxlen
#> Class:     numeric     numeric     numeric     numeric     numeric     numeric
#>                                                                               
#> Name:        isRel         LR5         LFR     Rmaxlen          DR    SelYears
#> Class:   character     numeric     numeric     numeric     numeric     numeric
#>                                                                               
#> Name:  AbsSelYears     L5Lower     L5Upper    LFSLower    LFSUpper   VmaxLower
#> Class:     numeric     numeric     numeric     numeric     numeric     numeric
#>                                           
#> Name:    VmaxUpper   CurrentYr         MPA
#> Class:     numeric     numeric      matrix
#> 
#> Known Subclasses: "OM"