Class 'Imp'

An operating model component that specifies the degree of adherence to management recommendations (Implementation error)

Slots

Name

The name of the Implementation error object. Single value. Character string.

TACFrac

Mean fraction of TAC taken. Uniform distribution lower and upper bounds. Positive real number.

TACSD

Log-normal coefficient of variation in the fraction of Total Allowable Catch (TAC) taken. Uniform distribution lower and upper bounds. Non-negative real numbers.

TAEFrac

Mean fraction of TAE taken. Uniform distribution lower and upper bounds. Positive real number.

TAESD

Log-normal coefficient of variation in the fraction of Total Allowable Effort (TAE) taken. Uniform distribution lower and upper bounds. Non-negative real numbers.

SizeLimFrac

The real minimum size that is retained expressed as a fraction of the size. Uniform distribution lower and upper bounds. Positive real number.

SizeLimSD

Log-normal coefficient of variation controlling mismatch between a minimum size limit and the real minimum size retained. Uniform distribution lower and upper bounds. Non-negative real numbers.

Objects from the Class

Objects can be created by calls of the form new('Imp')

Examples

showClass('Imp')

#> Class "Imp" [package "DLMtool"]
#> 
#> Slots:
#>                                                                               
#> Name:         Name     TACFrac       TACSD     TAEFrac       TAESD SizeLimFrac
#> Class:   character     numeric     numeric     numeric     numeric     numeric
#>                   
#> Name:    SizeLimSD
#> Class:     numeric
#> 
#> Known Subclasses: "OM"