What management procedures need what data

A function that finds all the MPs and searches the function text for slots in the Data object

Required(funcs = NA, noCV = FALSE)

Arguments

funcs
A character vector of management procedures

noCV
Logical. Should the CV slots be left out?

Value

A matrix of MPs and their required data in terms of Data slotnames

See also

Examples

Required(c("DCAC", "AvC"))

#>    MP    
#> 19 "DCAC"
#> 1  "AvC" 
#>    Data                                                                                    
#> 19 "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"
#> 1  "Cat, LHYear, Year"                                                                     

Required() # For all MPs

#>       MP            
#>  [1,] "AvC"         
#>  [2,] "AvC_MLL"     
#>  [3,] "BK"          
#>  [4,] "BK_CC"       
#>  [5,] "BK_ML"       
#>  [6,] "CC1"         
#>  [7,] "CC2"         
#>  [8,] "CC3"         
#>  [9,] "CC4"         
#> [10,] "CC5"         
#> [11,] "CompSRA"     
#> [12,] "CompSRA4010" 
#> [13,] "curE"        
#> [14,] "curE75"      
#> [15,] "DAAC"        
#> [16,] "DBSRA"       
#> [17,] "DBSRA_40"    
#> [18,] "DBSRA4010"   
#> [19,] "DCAC"        
#> [20,] "DCAC_40"     
#> [21,] "DCAC_ML"     
#> [22,] "DCAC4010"    
#> [23,] "DCACs"       
#> [24,] "DD"          
#> [25,] "DD4010"      
#> [26,] "DDe"         
#> [27,] "DDe75"       
#> [28,] "DDes"        
#> [29,] "DepF"        
#> [30,] "DTe40"       
#> [31,] "DTe50"       
#> [32,] "DynF"        
#> [33,] "EtargetLopt" 
#> [34,] "Fadapt"      
#> [35,] "Fdem"        
#> [36,] "Fdem_CC"     
#> [37,] "Fdem_ML"     
#> [38,] "FMSYref"     
#> [39,] "FMSYref50"   
#> [40,] "FMSYref75"   
#> [41,] "Fratio"      
#> [42,] "Fratio_CC"   
#> [43,] "Fratio_ML"   
#> [44,] "Fratio4010"  
#> [45,] "GB_CC"       
#> [46,] "GB_slope"    
#> [47,] "GB_target"   
#> [48,] "Gcontrol"    
#> [49,] "HDAAC"       
#> [50,] "ICI2"        
#> [51,] "Iratio"      
#> [52,] "Islope1"     
#> [53,] "Islope4"     
#> [54,] "IT10"        
#> [55,] "IT5"         
#> [56,] "Itarget1"    
#> [57,] "Itarget1_MPA"
#> [58,] "Itarget4"    
#> [59,] "ItargetE1"   
#> [60,] "ItargetE4"   
#> [61,] "ITe10"       
#> [62,] "ITe5"        
#> [63,] "ITM"         
#> [64,] "L95target"   
#> [65,] "Lratio_BHI"  
#> [66,] "Lratio_BHI2" 
#> [67,] "LstepCC1"    
#> [68,] "LstepCC4"    
#> [69,] "LstepCE1"    
#> [70,] "LstepCE2"    
#> [71,] "Ltarget1"    
#> [72,] "Ltarget4"    
#> [73,] "LtargetE1"   
#> [74,] "LtargetE4"   
#> [75,] "matlenlim"   
#> [76,] "matlenlim2"  
#> [77,] "MCD"         
#> [78,] "MCD4010"     
#> [79,] "minlenLopt1" 
#> [80,] "MRnoreal"    
#> [81,] "MRreal"      
#> [82,] "NFref"       
#> [83,] "Rcontrol"    
#> [84,] "Rcontrol2"   
#> [85,] "SBT1"        
#> [86,] "SBT2"        
#> [87,] "slotlim"     
#> [88,] "SPmod"       
#> [89,] "SPMSY"       
#> [90,] "SPslope"     
#> [91,] "SPSRA"       
#> [92,] "SPSRA_ML"    
#> [93,] "YPR"         
#> [94,] "YPR_CC"      
#> [95,] "YPR_ML"      
#> [96,] "ICI"         
#>       Data                                                                                                                                                                                                                                                                                                                                                                                     
#>  [1,] "Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                                      
#>  [2,] "Cat, L50"                                                                                                                                                                                                                                                                                                                                                                               
#>  [3,] "Abun, CV_Abun, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                                     
#>  [4,] "CAA, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, LFC, Mort, vbK, vbLinf"                                                                                                                                                                                                                                                                                                                   
#>  [5,] "CAL, CAL_bins, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, LFC, LFS, Mort, vbK, vbLinf"                                                                                                                                                                                                                                                                                                    
#>  [6,] "Cat, CV_Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                              
#>  [7,] "Cat, CV_Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                              
#>  [8,] "Cat, CV_Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                              
#>  [9,] "Cat, CV_Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                              
#> [10,] "Cat, CV_Cat, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                              
#> [11,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                       
#> [12,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                       
#> [13,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [14,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [15,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                 
#> [16,] "Cat, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                            
#> [17,] "Cat, CV_Dep, CV_FMSY_M, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                 
#> [18,] "Cat, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                            
#> [19,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                 
#> [20,] "AvC, BMSY_B0, CV_BMSY_B0, CV_FMSY_M, CV_Mort, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                            
#> [21,] "AvC, BMSY_B0, CAL, CAL_bins, Cat, CV_BMSY_B0, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, LHYear, Mort, t, vbK, vbLinf, Year"                                                                                                                                                                                                                                                   
#> [22,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                 
#> [23,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                 
#> [24,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                         
#> [25,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                         
#> [26,] "Cat, Ind, L50, LHYear, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                 
#> [27,] "Cat, Ind, L50, LHYear, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                                 
#> [28,] "Cat, Ind, L50, LHYear, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb, Year"                                                                                                                                                                                                                                                                                                          
#> [29,] "Abun, BMSY_B0, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort, Year"                                                                                                                                                                                                                                                                                                                     
#> [30,] "Dep, MPeff"                                                                                                                                                                                                                                                                                                                                                                             
#> [31,] "Dep, MPeff"                                                                                                                                                                                                                                                                                                                                                                             
#> [32,] "Abun, Cat, CV_FMSY_M, CV_Mort, FMSY_M, Ind, Mort, Year"                                                                                                                                                                                                                                                                                                                                 
#> [33,] "L50, ML, Mort, MPeff, vbK, vbLinf, wlb, Year"                                                                                                                                                                                                                                                                                                                                           
#> [34,] "Abun, Cat, FMSY_M, Ind, Mort, Year"                                                                                                                                                                                                                                                                                                                                                     
#> [35,] "Abun, CV_Abun, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                         
#> [36,] "CAA, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                      
#> [37,] "CAL, CAL_bins, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                       
#> [38,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [39,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [40,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [41,] "Abun, CV_Abun, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                                                                                                                                                                                                                                                        
#> [42,] "CAA, Cat, CV_Cat, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                                                                                                                                                                                                                                                     
#> [43,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                                                                                                                                                                                                                                                                      
#> [44,] "Abun, CV_Abun, CV_Dt, CV_FMSY_M, CV_Mort, Dep, Dt, FMSY_M, Mort"                                                                                                                                                                                                                                                                                                                        
#> [45,] "Cat, Cref, CV_Cref"                                                                                                                                                                                                                                                                                                                                                                     
#> [46,] "Cat, CV_Cat, Ind, Year"                                                                                                                                                                                                                                                                                                                                                                 
#> [47,] "Cat, Cref, CV_Cref, CV_Iref, Ind, Iref"                                                                                                                                                                                                                                                                                                                                                 
#> [48,] "Abun, Cat, Ind, Year"                                                                                                                                                                                                                                                                                                                                                                   
#> [49,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                                                                                                                                                                                                                                                 
#> [50,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                                                                                                                                                                                                                                                               
#> [51,] "Cat, CV_Cat, CV_Ind, Ind, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [52,] "Cat, CV_Cat, Ind, LHYear, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                  
#> [53,] "Cat, CV_Cat, Ind, LHYear, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                  
#> [54,] "CV_Ind, Ind, Iref, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [55,] "CV_Ind, Ind, Iref, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [56,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [57,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [58,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [59,] "Ind, LHYear, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                               
#> [60,] "Ind, LHYear, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                               
#> [61,] "CV_Ind, Ind, Iref, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [62,] "CV_Ind, Ind, Iref, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                         
#> [63,] "Cat, CV_Ind, Ind, Iref, Mort, MPrec, Year"                                                                                                                                                                                                                                                                                                                                              
#> [64,] "Cat, CV_Cat, L50, L95, LHYear, ML, Year"                                                                                                                                                                                                                                                                                                                                                
#> [65,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_vbLinf, LFS, vbLinf"                                                                                                                                                                                                                                                                                                                             
#> [66,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                                                                                                                                                                                                                                                              
#> [67,] "Cat, CV_Cat, LHYear, ML, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                   
#> [68,] "Cat, CV_Cat, LHYear, ML, MPrec, Year"                                                                                                                                                                                                                                                                                                                                                   
#> [69,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                
#> [70,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                
#> [71,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                                                                                                                                                                                                                                                          
#> [72,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                                                                                                                                                                                                                                                          
#> [73,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                
#> [74,] "LHYear, ML, MPeff, Year"                                                                                                                                                                                                                                                                                                                                                                
#> [75,] "L50"                                                                                                                                                                                                                                                                                                                                                                                    
#> [76,] "L50"                                                                                                                                                                                                                                                                                                                                                                                    
#> [77,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                                                                                                                                                                                                                                                               
#> [78,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                                                                                                                                                                                                                                                               
#> [79,] "Mort, MPeff, vbK, vbLinf, wlb"                                                                                                                                                                                                                                                                                                                                                          
#> [80,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [81,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [82,] ""                                                                                                                                                                                                                                                                                                                                                                                       
#> [83,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, Ind, MaxAge, Mort, steep, vbK, vbLinf, vbt0, Year"                                                                                                                                                                                                                                                               
#> [84,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, Ind, MaxAge, Mort, steep, vbK, vbLinf, vbt0, Year"                                                                                                                                                                                                                                                               
#> [85,] "Cat, CV_Cat, Ind, Year"                                                                                                                                                                                                                                                                                                                                                                 
#> [86,] "Cat, Cref, CV_Cref, Rec"                                                                                                                                                                                                                                                                                                                                                                
#> [87,] "L50, vbLinf"                                                                                                                                                                                                                                                                                                                                                                            
#> [88,] "Abun, Cat, CV_Abun, CV_Cat, CV_Ind, Ind"                                                                                                                                                                                                                                                                                                                                                
#> [89,] "Cat, L50, MaxAge, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                                                                                                    
#> [90,] "Abun, Cat, CV_Cat, Ind, Year"                                                                                                                                                                                                                                                                                                                                                           
#> [91,] "Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, MaxAge, Mort, steep, vbK, vbLinf, vbt0"                                                                                                                                                                                                                                                                        
#> [92,] "Abun, AvC, BMSY_B0, Bref, CAL, CAL_bins, Cat, Cref, CV_AvC, CV_BMSY_B0, CV_Bref, CV_Cat, CV_Cref, CV_Dep, CV_Dt, CV_FMSY_M, CV_Ind, CV_Iref, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_Rec, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, Dep, Dt, FMSY_M, Ind, Iref, L50, L95, Lbar, Lc, LenCV, LFS, MaxAge, ML, Mort, MPeff, Rec, sigmaL, SpAbun, steep, t, vbK, vbLinf, vbt0, Year"
#> [93,] "Abun, CV_Abun, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                             
#> [94,] "CAA, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                          
#> [95,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                                                                                                                                                                                                
#> [96,] "Cat, Ind, CV_Cat, CV_Ind"