What management procedures need what data

A function that finds all the MPs and searches the function text for slots in the Data object

Required(funcs = NA)

Arguments

funcs
A character vector of management procedures

Value

A matrix of MPs and their required data in terms of Data slotnames

See also

Examples

Required(c("DCAC", "AvC"))

#>      [,1]  
#> [1,] "DCAC"
#> [2,] "AvC" 
#>      [,2]                                                                      
#> [1,] "t, AvC, Dt, Mort, FMSY_M, BMSY_B0, CV_Dt, CV_Mort, CV_FMSY_M, CV_BMSY_B0"
#> [2,] "Year, Cat, Units, LHYear"                                                

Required() # For all MPs

#>       [,1]          
#>  [1,] "AvC"         
#>  [2,] "AvC_MLL"     
#>  [3,] "BK"          
#>  [4,] "BK_CC"       
#>  [5,] "BK_ML"       
#>  [6,] "CC1"         
#>  [7,] "CC4"         
#>  [8,] "CompSRA"     
#>  [9,] "CompSRA4010" 
#> [10,] "curE"        
#> [11,] "curE75"      
#> [12,] "DAAC"        
#> [13,] "DBSRA"       
#> [14,] "DBSRA_40"    
#> [15,] "DBSRA4010"   
#> [16,] "DCAC"        
#> [17,] "DCAC_40"     
#> [18,] "DCAC_ML"     
#> [19,] "DCAC4010"    
#> [20,] "DD"          
#> [21,] "DD4010"      
#> [22,] "DDe"         
#> [23,] "DDe75"       
#> [24,] "DDes"        
#> [25,] "DepF"        
#> [26,] "DTe40"       
#> [27,] "DTe50"       
#> [28,] "DynF"        
#> [29,] "EtargetLopt" 
#> [30,] "Fadapt"      
#> [31,] "Fdem"        
#> [32,] "Fdem_CC"     
#> [33,] "Fdem_ML"     
#> [34,] "FMSYref"     
#> [35,] "FMSYref50"   
#> [36,] "FMSYref75"   
#> [37,] "Fratio"      
#> [38,] "Fratio_CC"   
#> [39,] "Fratio_ML"   
#> [40,] "Fratio4010"  
#> [41,] "GB_CC"       
#> [42,] "GB_slope"    
#> [43,] "GB_target"   
#> [44,] "Gcontrol"    
#> [45,] "HDAAC"       
#> [46,] "ICI"         
#> [47,] "ICI2"        
#> [48,] "Iratio"      
#> [49,] "Islope1"     
#> [50,] "Islope4"     
#> [51,] "IT10"        
#> [52,] "IT5"         
#> [53,] "Itarget1"    
#> [54,] "Itarget1_MPA"
#> [55,] "Itarget4"    
#> [56,] "ItargetE1"   
#> [57,] "ItargetE4"   
#> [58,] "ITe10"       
#> [59,] "ITe5"        
#> [60,] "ITM"         
#> [61,] "L95target"   
#> [62,] "Lratio_BHI"  
#> [63,] "Lratio_BHI2" 
#> [64,] "LstepCC1"    
#> [65,] "LstepCC4"    
#> [66,] "LstepCE1"    
#> [67,] "LstepCE2"    
#> [68,] "Ltarget1"    
#> [69,] "Ltarget4"    
#> [70,] "LtargetE1"   
#> [71,] "LtargetE4"   
#> [72,] "matlenlim"   
#> [73,] "matlenlim2"  
#> [74,] "MCD"         
#> [75,] "MCD4010"     
#> [76,] "minlenLopt1" 
#> [77,] "MRnoreal"    
#> [78,] "MRreal"      
#> [79,] "NFref"       
#> [80,] "Rcontrol"    
#> [81,] "Rcontrol2"   
#> [82,] "SBT1"        
#> [83,] "SBT2"        
#> [84,] "slotlim"     
#> [85,] "SPmod"       
#> [86,] "SPMSY"       
#> [87,] "SPslope"     
#> [88,] "SPSRA"       
#> [89,] "SPSRA_ML"    
#> [90,] "YPR"         
#> [91,] "YPR_CC"      
#> [92,] "YPR_ML"      
#>       [,2]                                                                                                                                                              
#>  [1,] "Year, Cat, Units, LHYear"                                                                                                                                        
#>  [2,] "Cat, L50"                                                                                                                                                        
#>  [3,] "LFC, Abun, vbK, vbLinf, CV_Abun, CV_vbK, CV_vbLinf"                                                                                                              
#>  [4,] "Cat, Mort, LFC, CAA, vbK, vbLinf, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf"                                                                                            
#>  [5,] "Cat, Mort, LFC, vbK, vbLinf, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CAL"                                                                                            
#>  [6,] "Year, Cat, CV_Cat"                                                                                                                                               
#>  [7,] "Year, Cat, CV_Cat"                                                                                                                                               
#>  [8,] "Cat, Mort, L50, LFC, LFS, CAA, vbK, vbLinf, vbt0, wla, wlb, steep, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_L50, CV_LFC, CV_LFS, CV_wla, CV_wlb, CV_steep, MaxAge"
#>  [9,] "Cat, Mort, L50, LFC, LFS, CAA, vbK, vbLinf, vbt0, wla, wlb, steep, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_L50, CV_LFC, CV_LFS, CV_wla, CV_wlb, CV_steep, MaxAge"
#> [10,] ""                                                                                                                                                                
#> [11,] ""                                                                                                                                                                
#> [12,] "t, AvC, Dt, Mort, FMSY_M, BMSY_B0, CV_Dt, CV_Mort, CV_FMSY_M, CV_BMSY_B0"                                                                                        
#> [13,] "Cat, Mort, FMSY_M, BMSY_B0, L50, Dep, vbK, vbLinf, vbt0, CV_Mort, CV_FMSY_M, CV_BMSY_B0, CV_Dep"                                                                 
#> [14,] "Cat, Mort, FMSY_M, BMSY_B0, L50, Dep, vbK, vbLinf, vbt0, CV_Mort, CV_FMSY_M, CV_BMSY_B0, CV_Dep"                                                                 
#> [15,] "Cat, Mort, FMSY_M, BMSY_B0, L50, Dep, vbK, vbLinf, vbt0, CV_Mort, CV_FMSY_M, CV_BMSY_B0, CV_Dep"                                                                 
#> [16,] "t, AvC, Dt, Mort, FMSY_M, BMSY_B0, CV_Dt, CV_Mort, CV_FMSY_M, CV_BMSY_B0"                                                                                        
#> [17,] "t, AvC, Mort, FMSY_M, BMSY_B0, CV_Mort, CV_FMSY_M, CV_BMSY_B0"                                                                                                   
#> [18,] "Year, Cat, t, AvC, Mort, FMSY_M, BMSY_B0, vbK, vbLinf, CV_Mort, CV_FMSY_M, CV_BMSY_B0, CV_vbK, CV_vbLinf, CAL"                                                   
#> [19,] "t, AvC, Dt, Mort, FMSY_M, BMSY_B0, CV_Dt, CV_Mort, CV_FMSY_M, CV_BMSY_B0"                                                                                        
#> [20,] "Year, Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge"                                                                                                  
#> [21,] "Year, Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge"                                                                                                  
#> [22,] "Year, Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge, LHYear"                                                                                          
#> [23,] "Year, Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge, LHYear"                                                                                          
#> [24,] "Year, Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge, MPeff, LHYear"                                                                                   
#> [25,] "Year, Mort, FMSY_M, BMSY_B0, Dep, Abun, CV_Mort, CV_FMSY_M, CV_Dep"                                                                                              
#> [26,] "Dep, MPeff"                                                                                                                                                      
#> [27,] "Dep, MPeff"                                                                                                                                                      
#> [28,] "Year, Cat, Ind, Mort, FMSY_M, Abun, CV_Mort, CV_FMSY_M"                                                                                                          
#> [29,] "Year, Mort, L50, ML, vbK, vbLinf, wlb, MPeff"                                                                                                                    
#> [30,] "Year, Cat, Ind, Mort, FMSY_M, Abun"                                                                                                                              
#> [31,] "Mort, Abun, vbK, vbLinf, vbt0, wla, wlb, steep, CV_Mort, CV_Abun, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge"                                                  
#> [32,] "Cat, Mort, CAA, vbK, vbLinf, vbt0, wla, wlb, steep, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge"                                               
#> [33,] "Cat, Mort, vbK, vbLinf, vbt0, wla, wlb, steep, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge, CAL"                                               
#> [34,] "OM"                                                                                                                                                              
#> [35,] "OM"                                                                                                                                                              
#> [36,] "OM"                                                                                                                                                              
#> [37,] "Mort, FMSY_M, Abun, CV_Mort, CV_FMSY_M, CV_Abun"                                                                                                                 
#> [38,] "Cat, Mort, FMSY_M, CAA, CV_Cat, CV_Mort, CV_FMSY_M"                                                                                                              
#> [39,] "Cat, Mort, FMSY_M, vbK, vbLinf, CV_Cat, CV_Mort, CV_FMSY_M, CV_vbK, CV_vbLinf, CAL"                                                                              
#> [40,] "Dt, Mort, FMSY_M, Dep, Abun, CV_Dt, CV_Mort, CV_FMSY_M, CV_Abun"                                                                                                 
#> [41,] "Cat, Cref, CV_Cref"                                                                                                                                              
#> [42,] "Year, Cat, Ind, CV_Cat"                                                                                                                                          
#> [43,] "Cat, Ind, Cref, Iref, CV_Cref, CV_Iref"                                                                                                                          
#> [44,] "Year, Cat, Ind, Abun"                                                                                                                                            
#> [45,] "t, AvC, Dt, Mort, FMSY_M, BMSY_B0, CV_Dt, CV_Mort, CV_FMSY_M, CV_BMSY_B0"                                                                                        
#> [46,] "Cat, Ind, CV_Cat, CV_Ind"                                                                                                                                        
#> [47,] "Cat, Ind, CV_Cat, CV_Ind"                                                                                                                                        
#> [48,] "Year, Cat, Ind, CV_Cat, CV_Ind"                                                                                                                                  
#> [49,] "Year, Cat, Ind, CV_Cat, MPrec, LHYear"                                                                                                                           
#> [50,] "Year, Cat, Ind, CV_Cat, MPrec, LHYear"                                                                                                                           
#> [51,] "Year, Ind, CV_Ind, Iref, MPrec"                                                                                                                                  
#> [52,] "Year, Ind, CV_Ind, Iref, MPrec"                                                                                                                                  
#> [53,] "Year, Cat, Ind, CV_Cat, LHYear"                                                                                                                                  
#> [54,] "Year, Cat, Ind, CV_Cat, LHYear, nareas"                                                                                                                          
#> [55,] "Year, Cat, Ind, CV_Cat, LHYear"                                                                                                                                  
#> [56,] "Year, Ind, MPeff, LHYear"                                                                                                                                        
#> [57,] "Year, Ind, MPeff, LHYear"                                                                                                                                        
#> [58,] "Year, Ind, CV_Ind, Iref, MPeff"                                                                                                                                  
#> [59,] "Year, Ind, CV_Ind, Iref, MPeff"                                                                                                                                  
#> [60,] "Year, Cat, Ind, Mort, CV_Ind, Iref, MPrec"                                                                                                                       
#> [61,] "Year, Cat, L50, L95, ML, CV_Cat, LHYear"                                                                                                                         
#> [62,] "Cat, LFS, vbLinf, CV_Cat, CV_vbLinf, CV_LFS, CAL_bins, CAL"                                                                                                      
#> [63,] "Cat, Mort, FMSY_M, LFS, vbK, vbLinf, CV_Cat, CV_Mort, CV_FMSY_M, CV_vbK, CV_vbLinf, CV_LFS, CAL_bins, CAL"                                                       
#> [64,] "Year, Cat, ML, CV_Cat, MPrec, LHYear"                                                                                                                            
#> [65,] "Year, Cat, ML, CV_Cat, MPrec, LHYear"                                                                                                                            
#> [66,] "Year, ML, MPeff, LHYear"                                                                                                                                         
#> [67,] "Year, ML, MPeff, LHYear"                                                                                                                                         
#> [68,] "Year, Cat, ML, CV_Cat, LHYear"                                                                                                                                   
#> [69,] "Year, Cat, ML, CV_Cat, LHYear"                                                                                                                                   
#> [70,] "Year, ML, MPeff, LHYear"                                                                                                                                         
#> [71,] "Year, ML, MPeff, LHYear"                                                                                                                                         
#> [72,] "L50"                                                                                                                                                             
#> [73,] "L50"                                                                                                                                                             
#> [74,] "Cat, Dep, CV_Cat, CV_Dep"                                                                                                                                        
#> [75,] "Cat, Dep, CV_Cat, CV_Dep"                                                                                                                                        
#> [76,] "Mort, vbK, vbLinf, wlb, MPeff"                                                                                                                                   
#> [77,] "nareas"                                                                                                                                                          
#> [78,] "nareas"                                                                                                                                                          
#> [79,] ""                                                                                                                                                                
#> [80,] "Year, Cat, Ind, Mort, Dep, Abun, vbK, vbLinf, vbt0, steep, CV_Mort, CV_Dep, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge"                                        
#> [81,] "Year, Cat, Ind, Mort, Dep, Abun, vbK, vbLinf, vbt0, steep, CV_Mort, CV_Dep, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge"                                        
#> [82,] "Year, Cat, Ind, CV_Cat"                                                                                                                                          
#> [83,] "Cat, Rec, Cref, CV_Cref"                                                                                                                                         
#> [84,] "L50, vbLinf"                                                                                                                                                     
#> [85,] "Cat, Ind, Abun, CV_Cat, CV_Ind, CV_Abun"                                                                                                                         
#> [86,] "Cat, L50, vbK, vbLinf, vbt0, MaxAge"                                                                                                                             
#> [87,] "Year, Cat, Ind, Abun, CV_Cat"                                                                                                                                    
#> [88,] "Cat, Mort, Dep, vbK, vbLinf, vbt0, steep, CV_Cat, CV_Mort, CV_Dep, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge"                                                 
#> [89,] "Year, Cat, Mort, vbK, vbLinf, vbt0, steep, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_steep, MaxAge, CAL"                                                   
#> [90,] "Mort, LFS, Abun, vbK, vbLinf, vbt0, wla, wlb, CV_Mort, CV_Abun, CV_vbK, CV_vbLinf, CV_vbt0, CV_LFS, MaxAge"                                                      
#> [91,] "Cat, Mort, LFS, CAA, vbK, vbLinf, vbt0, wla, wlb, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_LFS, MaxAge"                                                   
#> [92,] "Cat, Mort, LFS, vbK, vbLinf, vbt0, wla, wlb, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_LFS, MaxAge, CAL"