What management procedures need what data

A function that finds all the MPs and searches the function text for slots in the Data object

Required(funcs = NA, noCV = FALSE)

Arguments

funcs
A character vector of management procedures

noCV
Logical. Should the CV slots be left out?

Value

A matrix of MPs and their required data in terms of Data slotnames

See also

Examples

Required(c("DCAC", "AvC"))

#>    MP    
#> 19 "DCAC"
#> 1  "AvC" 
#>    Data                                                                                    
#> 19 "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"
#> 1  "Cat, LHYear, Year"                                                                     

Required() # For all MPs

#>        MP            
#>   [1,] "AvC"         
#>   [2,] "AvC_MLL"     
#>   [3,] "BK"          
#>   [4,] "BK_CC"       
#>   [5,] "BK_ML"       
#>   [6,] "CC1"         
#>   [7,] "CC2"         
#>   [8,] "CC3"         
#>   [9,] "CC4"         
#>  [10,] "CC5"         
#>  [11,] "CompSRA"     
#>  [12,] "CompSRA4010" 
#>  [13,] "curE"        
#>  [14,] "curE75"      
#>  [15,] "DAAC"        
#>  [16,] "DBSRA"       
#>  [17,] "DBSRA_40"    
#>  [18,] "DBSRA4010"   
#>  [19,] "DCAC"        
#>  [20,] "DCAC_40"     
#>  [21,] "DCAC_ML"     
#>  [22,] "DCAC4010"    
#>  [23,] "DCACs"       
#>  [24,] "DD"          
#>  [25,] "DD4010"      
#>  [26,] "DDe"         
#>  [27,] "DDe75"       
#>  [28,] "DDes"        
#>  [29,] "DepF"        
#>  [30,] "DTe40"       
#>  [31,] "DTe50"       
#>  [32,] "DynF"        
#>  [33,] "EtargetLopt" 
#>  [34,] "Fadapt"      
#>  [35,] "Fdem"        
#>  [36,] "Fdem_CC"     
#>  [37,] "Fdem_ML"     
#>  [38,] "FMSYref"     
#>  [39,] "FMSYref50"   
#>  [40,] "FMSYref75"   
#>  [41,] "Fratio"      
#>  [42,] "Fratio_CC"   
#>  [43,] "Fratio_ML"   
#>  [44,] "Fratio4010"  
#>  [45,] "GB_CC"       
#>  [46,] "GB_slope"    
#>  [47,] "GB_target"   
#>  [48,] "Gcontrol"    
#>  [49,] "HDAAC"       
#>  [50,] "ICI"         
#>  [51,] "ICI2"        
#>  [52,] "Iratio"      
#>  [53,] "Islope1"     
#>  [54,] "Islope2"     
#>  [55,] "Islope4"     
#>  [56,] "IT10"        
#>  [57,] "IT5"         
#>  [58,] "Itarget1"    
#>  [59,] "Itarget1_MPA"
#>  [60,] "Itarget2"    
#>  [61,] "Itarget3"    
#>  [62,] "Itarget4"    
#>  [63,] "ItargetE1"   
#>  [64,] "ItargetE2"   
#>  [65,] "ItargetE3"   
#>  [66,] "ItargetE4"   
#>  [67,] "ITe10"       
#>  [68,] "ITe5"        
#>  [69,] "ITM"         
#>  [70,] "L95target"   
#>  [71,] "LBSPR"       
#>  [72,] "LBSPR_MLL"   
#>  [73,] "Lratio_BHI"  
#>  [74,] "Lratio_BHI2" 
#>  [75,] "Lratio_BHI3" 
#>  [76,] "LstepCC1"    
#>  [77,] "LstepCC2"    
#>  [78,] "LstepCC3"    
#>  [79,] "LstepCC4"    
#>  [80,] "LstepCE1"    
#>  [81,] "LstepCE2"    
#>  [82,] "Ltarget1"    
#>  [83,] "Ltarget2"    
#>  [84,] "Ltarget3"    
#>  [85,] "Ltarget4"    
#>  [86,] "LtargetE1"   
#>  [87,] "LtargetE4"   
#>  [88,] "matlenlim"   
#>  [89,] "matlenlim2"  
#>  [90,] "MCD"         
#>  [91,] "MCD4010"     
#>  [92,] "minlenLopt1" 
#>  [93,] "MRnoreal"    
#>  [94,] "MRreal"      
#>  [95,] "NFref"       
#>  [96,] "Rcontrol"    
#>  [97,] "Rcontrol2"   
#>  [98,] "SBT1"        
#>  [99,] "SBT2"        
#> [100,] "slotlim"     
#> [101,] "SPmod"       
#> [102,] "SPMSY"       
#> [103,] "SPslope"     
#> [104,] "SPSRA"       
#> [105,] "SPSRA_ML"    
#> [106,] "YPR"         
#> [107,] "YPR_CC"      
#> [108,] "YPR_ML"      
#>        Data                                                                                                                                                    
#>   [1,] "Cat, LHYear, Year"                                                                                                                                     
#>   [2,] "Cat, LHYear, Year"                                                                                                                                     
#>   [3,] "Abun, CV_Abun, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                    
#>   [4,] "CAA, Cat, CV_Cat, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                 
#>   [5,] "CAL, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, Lbar, Lc, LFC, Mort, vbK, vbLinf"                                                                        
#>   [6,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [7,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [8,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [9,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>  [10,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>  [11,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"
#>  [12,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"
#>  [13,] ""                                                                                                                                                      
#>  [14,] ""                                                                                                                                                      
#>  [15,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [16,] "BMSY_B0, Cat, CV_BMSY_B0, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                      
#>  [17,] "BMSY_B0, Cat, CV_BMSY_B0, CV_FMSY_M, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                                   
#>  [18,] "BMSY_B0, Cat, CV_BMSY_B0, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                      
#>  [19,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [20,] "AvC, BMSY_B0, CV_BMSY_B0, CV_FMSY_M, CV_Mort, FMSY_M, LHYear, Mort, t, Year"                                                                           
#>  [21,] "AvC, CAL, Cat, CV_Mort, CV_vbK, CV_vbLinf, Lbar, Lc, LHYear, Mort, t, vbK, vbLinf, Year"                                                               
#>  [22,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [23,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [24,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                              
#>  [25,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                              
#>  [26,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [27,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [28,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [29,] "Abun, CV_Abun, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort"                                                                                          
#>  [30,] "Dep, MPeff"                                                                                                                                            
#>  [31,] "Dep, MPeff"                                                                                                                                            
#>  [32,] "Abun, Cat, CV_FMSY_M, CV_Mort, FMSY_M, Ind, Mort, Year"                                                                                                
#>  [33,] "ML, Mort, MPeff, vbK, vbLinf, wlb"                                                                                                                     
#>  [34,] "Abun, Cat, FMSY_M, Ind, Mort, Year"                                                                                                                    
#>  [35,] "Abun, CV_Abun, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                           
#>  [36,] "CAA, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                        
#>  [37,] "CAL, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, Lbar, Lc, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"              
#>  [38,] ""                                                                                                                                                      
#>  [39,] ""                                                                                                                                                      
#>  [40,] ""                                                                                                                                                      
#>  [41,] "Abun, CV_Abun, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                       
#>  [42,] "CAA, Cat, CV_Cat, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                    
#>  [43,] "CAL, Cat, CV_Cat, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, Lbar, Lc, Mort, vbK, vbLinf"                                                          
#>  [44,] "Abun, CV_Abun, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort"                                                                                          
#>  [45,] "Cref, CV_Cref"                                                                                                                                         
#>  [46,] "Cat, CV_Cat, Ind, Year"                                                                                                                                
#>  [47,] "Cref, CV_Cref, CV_Iref, Ind, Iref"                                                                                                                     
#>  [48,] "Abun, Cat, Ind, Year"                                                                                                                                  
#>  [49,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [50,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [51,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [52,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [53,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [54,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [55,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [56,] "CV_Ind, Ind, Iref, MPrec"                                                                                                                              
#>  [57,] "CV_Ind, Ind, Iref, MPrec"                                                                                                                              
#>  [58,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [59,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [60,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [61,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [62,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [63,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [64,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [65,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [66,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [67,] "CV_Ind, Ind, Iref, MPeff"                                                                                                                              
#>  [68,] "CV_Ind, Ind, Iref, MPeff"                                                                                                                              
#>  [69,] "CV_Ind, Ind, Iref, Mort, MPrec"                                                                                                                        
#>  [70,] "Cat, CV_Cat, L50, LHYear, ML, Year"                                                                                                                    
#>  [71,] "CAL, CAL_bins, L50, L95, LHYear, Mort, vbK, vbLinf, wlb, Year"                                                                                         
#>  [72,] "CAL, CAL_bins, L50, L95, LHYear, Mort, vbK, vbLinf, wlb, Year"                                                                                         
#>  [73,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_vbLinf, LFS, vbLinf"                                                                                            
#>  [74,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                             
#>  [75,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                             
#>  [76,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [77,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [78,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [79,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [80,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [81,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [82,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [83,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [84,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [85,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [86,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [87,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [88,] "L50"                                                                                                                                                   
#>  [89,] "L50"                                                                                                                                                   
#>  [90,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                              
#>  [91,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                              
#>  [92,] "Mort, vbK, vbLinf, wlb"                                                                                                                                
#>  [93,] ""                                                                                                                                                      
#>  [94,] ""                                                                                                                                                      
#>  [95,] ""                                                                                                                                                      
#>  [96,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, Ind, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb, Year"       
#>  [97,] "Abun, Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, Ind, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb, Year"                    
#>  [98,] "Cat, CV_Cat, Ind, Year"                                                                                                                                
#>  [99,] "Cat, Cref, CV_Cref, Rec"                                                                                                                               
#> [100,] "L50, vbLinf"                                                                                                                                           
#> [101,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#> [102,] "Cat, L50, MaxAge, vbK, vbLinf, vbt0"                                                                                                                   
#> [103,] "Abun, Cat, CV_Cat, Ind, Year"                                                                                                                          
#> [104,] "Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                
#> [105,] "CAL, Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, L50, Lbar, Lc, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb" 
#> [106,] "Abun, CV_Abun, CV_LFS, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, vbK, vbLinf, vbt0"                                                                     
#> [107,] "CAA, Cat, CV_Cat, CV_LFS, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, vbK, vbLinf, vbt0"                                                                  
#> [108,] "CAL, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, Lbar, Lc, LFS, MaxAge, Mort, vbK, vbLinf, vbt0"