What management procedures need what data

A function that finds all the MPs and searches the function text for slots in the Data object

Required(funcs = NA, noCV = FALSE)

Arguments

funcs
A character vector of management procedures

noCV
Logical. Should the CV slots be left out?

Value

A matrix of MPs and their required data in terms of slotnames('Data'), and broad Data classes for each MP

See also

Examples

Required(c("DCAC", "AvC"))

#>    MP    
#> 20 "DCAC"
#> 1  "AvC" 
#>    Data                                                                                    
#> 20 "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"
#> 1  "Cat, LHYear, Year"                                                                     
#>    DataClass                                                              
#> 20 "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"
#> 1  "Recent Catches"                                                       

Required() # For all MPs

#>        MP            
#>   [1,] "AvC"         
#>   [2,] "AvC_MLL"     
#>   [3,] "BK"          
#>   [4,] "BK_CC"       
#>   [5,] "BK_ML"       
#>   [6,] "CC1"         
#>   [7,] "CC2"         
#>   [8,] "CC3"         
#>   [9,] "CC4"         
#>  [10,] "CC5"         
#>  [11,] "CompSRA"     
#>  [12,] "CompSRA4010" 
#>  [13,] "CurC"        
#>  [14,] "curE"        
#>  [15,] "curE75"      
#>  [16,] "DAAC"        
#>  [17,] "DBSRA"       
#>  [18,] "DBSRA_40"    
#>  [19,] "DBSRA4010"   
#>  [20,] "DCAC"        
#>  [21,] "DCAC_40"     
#>  [22,] "DCAC_ML"     
#>  [23,] "DCAC4010"    
#>  [24,] "DCACs"       
#>  [25,] "DD"          
#>  [26,] "DD4010"      
#>  [27,] "DDe"         
#>  [28,] "DDe75"       
#>  [29,] "DDes"        
#>  [30,] "DepF"        
#>  [31,] "DTe40"       
#>  [32,] "DTe50"       
#>  [33,] "DynF"        
#>  [34,] "EtargetLopt" 
#>  [35,] "Fadapt"      
#>  [36,] "Fdem"        
#>  [37,] "Fdem_CC"     
#>  [38,] "Fdem_ML"     
#>  [39,] "FMSYref"     
#>  [40,] "FMSYref50"   
#>  [41,] "FMSYref75"   
#>  [42,] "Fratio"      
#>  [43,] "Fratio_CC"   
#>  [44,] "Fratio_ML"   
#>  [45,] "Fratio4010"  
#>  [46,] "GB_CC"       
#>  [47,] "GB_slope"    
#>  [48,] "GB_target"   
#>  [49,] "Gcontrol"    
#>  [50,] "HDAAC"       
#>  [51,] "ICI"         
#>  [52,] "ICI2"        
#>  [53,] "Iratio"      
#>  [54,] "Islope1"     
#>  [55,] "Islope2"     
#>  [56,] "Islope3"     
#>  [57,] "Islope4"     
#>  [58,] "IT10"        
#>  [59,] "IT5"         
#>  [60,] "Itarget1"    
#>  [61,] "Itarget1_MPA"
#>  [62,] "Itarget2"    
#>  [63,] "Itarget3"    
#>  [64,] "Itarget4"    
#>  [65,] "ItargetE1"   
#>  [66,] "ItargetE2"   
#>  [67,] "ItargetE3"   
#>  [68,] "ItargetE4"   
#>  [69,] "ITe10"       
#>  [70,] "ITe5"        
#>  [71,] "ITM"         
#>  [72,] "L95target"   
#>  [73,] "LBSPR"       
#>  [74,] "LBSPR_MLL"   
#>  [75,] "Lratio_BHI"  
#>  [76,] "Lratio_BHI2" 
#>  [77,] "Lratio_BHI3" 
#>  [78,] "LstepCC1"    
#>  [79,] "LstepCC2"    
#>  [80,] "LstepCC3"    
#>  [81,] "LstepCC4"    
#>  [82,] "LstepCE1"    
#>  [83,] "LstepCE2"    
#>  [84,] "Ltarget1"    
#>  [85,] "Ltarget2"    
#>  [86,] "Ltarget3"    
#>  [87,] "Ltarget4"    
#>  [88,] "LtargetE1"   
#>  [89,] "LtargetE4"   
#>  [90,] "matlenlim"   
#>  [91,] "matlenlim2"  
#>  [92,] "MCD"         
#>  [93,] "MCD4010"     
#>  [94,] "minlenLopt1" 
#>  [95,] "MRnoreal"    
#>  [96,] "MRreal"      
#>  [97,] "NFref"       
#>  [98,] "Rcontrol"    
#>  [99,] "Rcontrol2"   
#> [100,] "SBT1"        
#> [101,] "SBT2"        
#> [102,] "slotlim"     
#> [103,] "SPmod"       
#> [104,] "SPMSY"       
#> [105,] "SPslope"     
#> [106,] "SPSRA"       
#> [107,] "SPSRA_ML"    
#> [108,] "YPR"         
#> [109,] "YPR_CC"      
#> [110,] "YPR_ML"      
#>        Data                                                                                                                                                    
#>   [1,] "Cat, LHYear, Year"                                                                                                                                     
#>   [2,] "Cat, LHYear, Year"                                                                                                                                     
#>   [3,] "Abun, CV_Abun, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                    
#>   [4,] "CAA, Cat, CV_Cat, CV_vbK, CV_vbLinf, LFC, vbK, vbLinf"                                                                                                 
#>   [5,] "CAL, Cat, CV_Cat, CV_Mort, CV_vbK, CV_vbLinf, Lbar, Lc, LFC, Mort, vbK, vbLinf"                                                                        
#>   [6,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [7,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [8,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>   [9,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>  [10,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>  [11,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"
#>  [12,] "CAA, Cat, CV_L50, CV_LFC, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, CV_wla, CV_wlb, L50, LFC, LFS, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"
#>  [13,] "Cat, CV_Cat, LHYear, Year"                                                                                                                             
#>  [14,] ""                                                                                                                                                      
#>  [15,] ""                                                                                                                                                      
#>  [16,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [17,] "BMSY_B0, Cat, CV_BMSY_B0, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                      
#>  [18,] "BMSY_B0, Cat, CV_BMSY_B0, CV_FMSY_M, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                                   
#>  [19,] "BMSY_B0, Cat, CV_BMSY_B0, CV_Dep, CV_FMSY_M, Dep, FMSY_M, L50, vbK, vbLinf, vbt0"                                                                      
#>  [20,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [21,] "AvC, BMSY_B0, CV_BMSY_B0, CV_FMSY_M, CV_Mort, FMSY_M, LHYear, Mort, t, Year"                                                                           
#>  [22,] "AvC, CAL, Cat, CV_Mort, CV_vbK, CV_vbLinf, Lbar, Lc, LHYear, Mort, t, vbK, vbLinf, Year"                                                               
#>  [23,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [24,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [25,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                              
#>  [26,] "Cat, Ind, L50, MaxAge, Mort, vbK, vbLinf, vbt0, wla, wlb"                                                                                              
#>  [27,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [28,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [29,] "Cat, Ind, L50, MaxAge, Mort, MPeff, vbK, vbLinf, vbt0, wla, wlb"                                                                                       
#>  [30,] "Abun, CV_Abun, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort"                                                                                          
#>  [31,] "Dep, MPeff"                                                                                                                                            
#>  [32,] "Dep, MPeff"                                                                                                                                            
#>  [33,] "Abun, Cat, CV_FMSY_M, CV_Mort, FMSY_M, Ind, Mort, Year"                                                                                                
#>  [34,] "ML, Mort, MPeff, vbK, vbLinf, wlb"                                                                                                                     
#>  [35,] "Abun, Cat, FMSY_M, Ind, Mort, Year"                                                                                                                    
#>  [36,] "Abun, CV_Abun, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                           
#>  [37,] "CAA, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                        
#>  [38,] "CAL, Cat, CV_Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, L50, Lbar, Lc, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"              
#>  [39,] ""                                                                                                                                                      
#>  [40,] ""                                                                                                                                                      
#>  [41,] ""                                                                                                                                                      
#>  [42,] "Abun, CV_Abun, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                       
#>  [43,] "CAA, Cat, CV_Cat, CV_FMSY_M, CV_Mort, FMSY_M, Mort"                                                                                                    
#>  [44,] "CAL, Cat, CV_Cat, CV_FMSY_M, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, Lbar, Lc, Mort, vbK, vbLinf"                                                          
#>  [45,] "Abun, CV_Abun, CV_Dep, CV_FMSY_M, CV_Mort, Dep, FMSY_M, Mort"                                                                                          
#>  [46,] "Cref, CV_Cref"                                                                                                                                         
#>  [47,] "Cat, CV_Cat, Ind, Year"                                                                                                                                
#>  [48,] "Cref, CV_Cref, CV_Iref, Ind, Iref"                                                                                                                     
#>  [49,] "Abun, Cat, Ind, Year"                                                                                                                                  
#>  [50,] "AvC, BMSY_B0, CV_BMSY_B0, CV_Dt, CV_FMSY_M, CV_Mort, Dt, FMSY_M, LHYear, Mort, t, Year"                                                                
#>  [51,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [52,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [53,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#>  [54,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [55,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [56,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [57,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [58,] "CV_Ind, Ind, Iref, MPrec"                                                                                                                              
#>  [59,] "CV_Ind, Ind, Iref, MPrec"                                                                                                                              
#>  [60,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [61,] "Cat, Ind, LHYear, Year"                                                                                                                                
#>  [62,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [63,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [64,] "Cat, CV_Cat, Ind, LHYear, Year"                                                                                                                        
#>  [65,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [66,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [67,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [68,] "Ind, LHYear, MPeff, Year"                                                                                                                              
#>  [69,] "CV_Ind, Ind, Iref, MPeff"                                                                                                                              
#>  [70,] "CV_Ind, Ind, Iref, MPeff"                                                                                                                              
#>  [71,] "CV_Ind, Ind, Iref, Mort, MPrec"                                                                                                                        
#>  [72,] "Cat, CV_Cat, L50, LHYear, ML, Year"                                                                                                                    
#>  [73,] "CAL, CAL_bins, L50, L95, LHYear, Mort, vbK, vbLinf, wlb, Year"                                                                                         
#>  [74,] "CAL, CAL_bins, L50, L95, LHYear, Mort, vbK, vbLinf, wlb, Year"                                                                                         
#>  [75,] "CAL, CAL_bins, Cat, CV_Cat, CV_LFS, CV_vbLinf, LFS, vbLinf"                                                                                            
#>  [76,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                             
#>  [77,] "CAL, CAL_bins, Cat, CV_Cat, CV_FMSY_M, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, FMSY_M, LFS, Mort, vbK, vbLinf"                                             
#>  [78,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [79,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [80,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [81,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [82,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [83,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [84,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [85,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [86,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [87,] "Cat, CV_Cat, LHYear, ML, Year"                                                                                                                         
#>  [88,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [89,] "LHYear, ML, MPeff, Year"                                                                                                                               
#>  [90,] "L50"                                                                                                                                                   
#>  [91,] "L50"                                                                                                                                                   
#>  [92,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                              
#>  [93,] "Cat, CV_Cat, CV_Dep, Dep"                                                                                                                              
#>  [94,] "Mort, vbK, vbLinf, wlb"                                                                                                                                
#>  [95,] ""                                                                                                                                                      
#>  [96,] ""                                                                                                                                                      
#>  [97,] ""                                                                                                                                                      
#>  [98,] "Abun, Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, Ind, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb, Year"       
#>  [99,] "Abun, Cat, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, FMSY_M, Ind, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb, Year"                    
#> [100,] "Cat, CV_Cat, Ind, Year"                                                                                                                                
#> [101,] "Cat, Cref, CV_Cref, Rec"                                                                                                                               
#> [102,] "L50, vbLinf"                                                                                                                                           
#> [103,] "Cat, CV_Cat, CV_Ind, Ind"                                                                                                                              
#> [104,] "Cat, L50, MaxAge, vbK, vbLinf, vbt0"                                                                                                                   
#> [105,] "Abun, Cat, CV_Cat, Ind, Year"                                                                                                                          
#> [106,] "Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, L50, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb"                
#> [107,] "CAL, Cat, CV_Cat, CV_Dep, CV_Mort, CV_steep, CV_vbK, CV_vbLinf, CV_vbt0, Dep, FMSY_M, L50, Lbar, Lc, MaxAge, Mort, steep, vbK, vbLinf, vbt0, wla, wlb" 
#> [108,] "Abun, CV_Abun, CV_LFS, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, vbK, vbLinf, vbt0"                                                                     
#> [109,] "CAA, Cat, CV_Cat, CV_LFS, CV_vbK, CV_vbLinf, CV_vbt0, LFS, MaxAge, vbK, vbLinf, vbt0"                                                                  
#> [110,] "CAL, Cat, CV_Cat, CV_LFS, CV_Mort, CV_vbK, CV_vbLinf, CV_vbt0, Lbar, Lc, LFS, MaxAge, Mort, vbK, vbLinf, vbt0"                                         
#>        DataClass                                                                                                                                                                 
#>   [1,] "Recent Catches"                                                                                                                                                          
#>   [2,] "Recent Catches"                                                                                                                                                          
#>   [3,] "Current Abundance, Growth Parameters, Selectivity"                                                                                                                       
#>   [4,] "Age Composition, Growth Parameters, Historical Catches, Selectivity"                                                                                                     
#>   [5,] "Growth Parameters, Historical Catches, Length Composition, Mean Length, Natural Mortality, Selectivity"                                                                  
#>   [6,] "Recent Catches"                                                                                                                                                          
#>   [7,] "Recent Catches"                                                                                                                                                          
#>   [8,] "Recent Catches"                                                                                                                                                          
#>   [9,] "Recent Catches"                                                                                                                                                          
#>  [10,] "Recent Catches"                                                                                                                                                          
#>  [11,] "Age Composition, Growth Parameters, Historical Catches, Maturity, Natural Mortality, Selectivity, Stock-Recruitment"                                                     
#>  [12,] "Age Composition, Growth Parameters, Historical Catches, Maturity, Natural Mortality, Selectivity, Stock-Recruitment"                                                     
#>  [13,] "Recent Catches"                                                                                                                                                          
#>  [14,] ""                                                                                                                                                                        
#>  [15,] ""                                                                                                                                                                        
#>  [16,] "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"                                                                                                   
#>  [17,] "Current Depletion, Growth Parameters, Historical Catches, Maturity, Reference Ratios"                                                                                    
#>  [18,] "Growth Parameters, Historical Catches, Maturity, Reference Ratios"                                                                                                       
#>  [19,] "Current Depletion, Growth Parameters, Historical Catches, Maturity, Reference Ratios"                                                                                    
#>  [20,] "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"                                                                                                   
#>  [21,] "Average Catch, Natural Mortality, Reference Ratios"                                                                                                                      
#>  [22,] "Average Catch, Growth Parameters, Length Composition, Mean Length, Natural Mortality, Recent Catches, Selectivity"                                                       
#>  [23,] "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"                                                                                                   
#>  [24,] "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"                                                                                                   
#>  [25,] "Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality"                                                                                  
#>  [26,] "Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality"                                                                                  
#>  [27,] "Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality, Recent Management"                                                               
#>  [28,] "Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality, Recent Management"                                                               
#>  [29,] "Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality, Recent Management"                                                               
#>  [30,] "Current Abundance, Current Depletion, Natural Mortality, Reference Ratios"                                                                                               
#>  [31,] "Current Depletion, Recent Management"                                                                                                                                    
#>  [32,] "Current Depletion, Recent Management"                                                                                                                                    
#>  [33,] "Current Abundance, Historical Catches, Index of Abundance, Natural Mortality, Reference Ratios"                                                                          
#>  [34,] "Growth Parameters, Mean Length, Natural Mortality, Recent Management"                                                                                                    
#>  [35,] "Current Abundance, Historical Catches, Index of Abundance, Natural Mortality, Reference Ratios"                                                                          
#>  [36,] "Current Abundance, Growth Parameters, Maturity, Natural Mortality, Reference Ratios, Stock-Recruitment"                                                                  
#>  [37,] "Age Composition, Growth Parameters, Historical Catches, Maturity, Natural Mortality, Reference Ratios, Stock-Recruitment"                                                
#>  [38,] "Growth Parameters, Historical Catches, Length Composition, Maturity, Mean Length, Natural Mortality, Reference Ratios, Selectivity, Stock-Recruitment"                   
#>  [39,] ""                                                                                                                                                                        
#>  [40,] ""                                                                                                                                                                        
#>  [41,] ""                                                                                                                                                                        
#>  [42,] "Current Abundance, Natural Mortality, Reference Ratios"                                                                                                                  
#>  [43,] "Age Composition, Historical Catches, Natural Mortality, Reference Ratios"                                                                                                
#>  [44,] "Growth Parameters, Historical Catches, Length Composition, Mean Length, Natural Mortality, Reference Ratios, Selectivity"                                                
#>  [45,] "Current Abundance, Current Depletion, Natural Mortality, Reference Ratios"                                                                                               
#>  [46,] "Reference Levels"                                                                                                                                                        
#>  [47,] "Historical Catches, Index of Abundance"                                                                                                                                  
#>  [48,] "Index of Abundance, Reference Levels"                                                                                                                                    
#>  [49,] "Current Abundance, Historical Catches, Index of Abundance"                                                                                                               
#>  [50,] "Average Catch, Current Depletion, Natural Mortality, Reference Ratios"                                                                                                   
#>  [51,] "Historical Catches, Index of Abundance"                                                                                                                                  
#>  [52,] "Historical Catches, Index of Abundance"                                                                                                                                  
#>  [53,] "Historical Catches, Index of Abundance"                                                                                                                                  
#>  [54,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [55,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [56,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [57,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [58,] "Index of Abundance, Recent Management, Reference Levels"                                                                                                                 
#>  [59,] "Index of Abundance, Recent Management, Reference Levels"                                                                                                                 
#>  [60,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [61,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [62,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [63,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [64,] "Index of Abundance, Recent Catches"                                                                                                                                      
#>  [65,] "Index of Abundance, Recent Management"                                                                                                                                   
#>  [66,] "Index of Abundance, Recent Management"                                                                                                                                   
#>  [67,] "Index of Abundance, Recent Management"                                                                                                                                   
#>  [68,] "Index of Abundance, Recent Management"                                                                                                                                   
#>  [69,] "Index of Abundance, Recent Management, Reference Levels"                                                                                                                 
#>  [70,] "Index of Abundance, Recent Management, Reference Levels"                                                                                                                 
#>  [71,] "Index of Abundance, Natural Mortality, Recent Management, Reference Levels"                                                                                              
#>  [72,] "Maturity, Mean Length, Recent Catches"                                                                                                                                   
#>  [73,] "Growth Parameters, Length Composition, Maturity, Natural Mortality"                                                                                                      
#>  [74,] "Growth Parameters, Length Composition, Maturity, Natural Mortality"                                                                                                      
#>  [75,] "Growth Parameters, Historical Catches, Length Composition, Selectivity"                                                                                                  
#>  [76,] "Growth Parameters, Historical Catches, Length Composition, Natural Mortality, Reference Ratios, Selectivity"                                                             
#>  [77,] "Growth Parameters, Historical Catches, Length Composition, Natural Mortality, Reference Ratios, Selectivity"                                                             
#>  [78,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [79,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [80,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [81,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [82,] "Mean Length, Recent Management"                                                                                                                                          
#>  [83,] "Mean Length, Recent Management"                                                                                                                                          
#>  [84,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [85,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [86,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [87,] "Mean Length, Recent Catches"                                                                                                                                             
#>  [88,] "Mean Length, Recent Management"                                                                                                                                          
#>  [89,] "Mean Length, Recent Management"                                                                                                                                          
#>  [90,] "Maturity"                                                                                                                                                                
#>  [91,] "Maturity"                                                                                                                                                                
#>  [92,] "Current Depletion, Historical Catches"                                                                                                                                   
#>  [93,] "Current Depletion, Historical Catches"                                                                                                                                   
#>  [94,] "Growth Parameters, Natural Mortality"                                                                                                                                    
#>  [95,] ""                                                                                                                                                                        
#>  [96,] ""                                                                                                                                                                        
#>  [97,] ""                                                                                                                                                                        
#>  [98,] "Current Abundance, Current Depletion, Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality, Reference Ratios, Stock-Recruitment"       
#>  [99,] "Current Abundance, Growth Parameters, Historical Catches, Index of Abundance, Maturity, Natural Mortality, Reference Ratios, Stock-Recruitment"                          
#> [100,] "Historical Catches, Index of Abundance"                                                                                                                                  
#> [101,] "Historical Catches, Recruitment Index, Recruitment Index, Reference Levels"                                                                                              
#> [102,] "Growth Parameters, Maturity"                                                                                                                                             
#> [103,] "Historical Catches, Index of Abundance"                                                                                                                                  
#> [104,] "Growth Parameters, Historical Catches, Maturity, Natural Mortality"                                                                                                      
#> [105,] "Current Abundance, Historical Catches, Index of Abundance"                                                                                                               
#> [106,] "Current Depletion, Growth Parameters, Historical Catches, Maturity, Natural Mortality, Reference Ratios, Stock-Recruitment"                                              
#> [107,] "Current Depletion, Growth Parameters, Historical Catches, Length Composition, Maturity, Mean Length, Natural Mortality, Reference Ratios, Selectivity, Stock-Recruitment"
#> [108,] "Current Abundance, Growth Parameters, Natural Mortality, Selectivity"                                                                                                    
#> [109,] "Age Composition, Growth Parameters, Historical Catches, Natural Mortality, Selectivity"                                                                                  
#> [110,] "Growth Parameters, Historical Catches, Length Composition, Mean Length, Natural Mortality, Selectivity"