Generic Trade-Plot Function - IN DEVELOPMENT

Generic Trade-Plot Function - IN DEVELOPMENT

TradePlot(MSEobj, ..., lims = c(0.2, 0.2, 0.8, 0.8), point.size = 2,
  lab.size = 4, axis.title.size = 12, axis.text.size = 10,
  legend.title.size = 12, position = c("right", "bottom"),
  fill = "gray80", alpha = 0.4, PMlist = NULL, Refs = NULL,
  Yrs = NULL)

Tplot(MSEobj, lims = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5), ...)

Tplot2(MSEobj, lims = c(0.2, 0.2, 0.8, 0.8), ...)

Tplot3(MSEobj, lims = c(0.5, 0.5, 0.8, 0.5), ...)

Arguments

MSEobj
An object of class MSE

...
Names of Performance Metrics (PMs), or other arguments to TradePlot. First PM is recycled if number of PMs is not even

lims
A numeric vector of acceptable risk/minimum probability thresholds. Recycled if not equal to number of PMs.

point.size
Numeric. Size of the MP points

lab.size
Numeric. Size of MP label

axis.title.size
Numeric. Size of axis titles

axis.text.size
Numeric. Size of axis text

legend.title.size
Numeric. Size of legend title text

position
Character. Position of legend - 'right' or 'bottom'

fill
Character. Color of the fill

alpha
Numeric. Transparency of fill

PMlist
Optional list of PM names. Overrides any supplied in ... above

Refs
An optional named list (matching the PM names) with numeric values to override the default Ref values. See examples.

Yrs
An optional named list (matching the PM names) with numeric values to override the default Yrs values. See examples.

Value

A summary table of MP performance

Functions

Tplot: A trade-off plot showing probabilities that:

overfishing (POF) against long-term yield is > 50% of reference yield (LTY)

spawning biomass is below BMSY (P100) against LTY

spawning biomass is below 0.5BMSY (P50) against LTY

spawning biomass is below 0.1BMSY (P10) against LTY

Tplot2: A trade-off plot showing probabilities that:

short-term yield is > 50

spawning biomass is below 0.1BMSY (P10) against average annual variability in yield is < 20% (AAVY)

Tplot3: A trade-off plot showing probabilities that:

overfishing (POF) against long-term yield is > 50

spawning biomass is below 0.1BMSY (P10) against average annual variability in yield is < 20% (AAVY)