Run a Management Strategy Evaluation

A function that runs a Management Strategy Evaluation (closed-loop simulation) for a specified operating model

runMSE(OM = DLMtool::testOM, MPs = c("AvC", "DCAC", "FMSYref", "curE",
  "matlenlim", "MRreal"), CheckMPs = FALSE, timelimit = 1,
  Hist = FALSE, ntrials = 50, fracD = 0.05, CalcBlow = TRUE,
  HZN = 2, Bfrac = 0.5, AnnualMSY = TRUE, silent = FALSE,
  PPD = FALSE, parallel = FALSE, save_name = NULL, checks = FALSE,
  control = NULL)

runMSErobust(...)

Arguments

OM
An operating model object (class 'OM')

MPs
A vector of methods (character string) of class MP

CheckMPs
Logical to indicate if Can function should be used to check if MPs can be run.

timelimit
Maximum time taken for a method to carry out 10 reps (methods are ignored that take longer)

Hist
Should model stop after historical simulations? Returns an object of class 'Hist' containing all historical data

ntrials
Maximum of times depletion and recruitment deviations are resampled to optimize for depletion. After this the model stops if more than percent of simulations are not close to the required depletion

fracD
Maximum allowed proportion of simulations where depletion is not close to sampled depletion from OM before model stops with error

CalcBlow
Should low biomass be calculated where this is the spawning biomass at which it takes HZN mean generation times of zero fishing to reach Bfrac fraction of SSBMSY

HZN
The number of mean generation times required to reach Bfrac SSBMSY in the Blow calculation

Bfrac
The target fraction of SSBMSY for calculating Blow

AnnualMSY
Deprecated. Always set to TRUE now. Logical. Should MSY statistics be calculated for each projection year? May differ from MSY statistics from last historical year if there are changes in productivity

silent
Should messages be printed out to the console?

PPD
Logical. Should posterior predicted data be included in the MSE object Misc slot?

parallel
Logical. Should the MSE be run using parallel processing?

save_name
Character. Optional name to save parallel MSE list

checks
Logical. Run tests?

control
control options for testing and debugging

...
Arguments to runMSErobust function (function no longer works)

Value

An object of class MSE

Functions

runMSE: Default function to use.

runMSErobust: This function has now been removed. Use runMSE.

Note

See relevant section of the DLMtool User Guide for more information.

See also