Valid custom parameters (cpars)

Valid custom parameters (cpars)

validcpars(type = c("all", "Stock", "Fleet", "Obs", "Imp", "internal"),
  valid = TRUE)

Arguments

type
What cpars to show? 'all', 'Stock', 'Fleet', 'Obs', 'Imp', or 'internal'

valid
Logical. Show valid cpars?

Value

a dataframe with variable name, description and type of valid/invalid cpars

Examples

validcpars() # all valid cpars

#>            Var.                                                   Dim.
#> 1            R0                             numeric vector length nsim
#> 2             M                             numeric vector length nsim
#> 3          Mexp                             numeric vector length nsim
#> 4           Msd                             numeric vector length nsim
#> 5         Mgrad                             numeric vector length nsim
#> 6             h                             numeric vector length nsim
#> 7         SRrel                             numeric vector length nsim
#> 8          Perr                             numeric vector length nsim
#> 9            AC                             numeric vector length nsim
#> 10       Period                             numeric vector length nsim
#> 11    Amplitude                             numeric vector length nsim
#> 12         Linf                             numeric vector length nsim
#> 13            K                             numeric vector length nsim
#> 14           t0                             numeric vector length nsim
#> 15        LenCV                             numeric vector length nsim
#> 16          Ksd                             numeric vector length nsim
#> 17        Kgrad                             numeric vector length nsim
#> 18       Linfsd                             numeric vector length nsim
#> 19     Linfgrad                             numeric vector length nsim
#> 20          L50                             numeric vector length nsim
#> 21       L50_95                             numeric vector length nsim
#> 22            D                             numeric vector length nsim
#> 23  Size_area_1                             numeric vector length nsim
#> 24  Frac_area_1                             numeric vector length nsim
#> 25 Prob_staying                             numeric vector length nsim
#> 26        Fdisc                             numeric vector length nsim
#> 27    Spat_targ                             numeric vector length nsim
#> 28          Esd                             numeric vector length nsim
#> 29         qinc                             numeric vector length nsim
#> 30          qcv                             numeric vector length nsim
#> 31           L5                             numeric vector length nsim
#> 32          LFS                             numeric vector length nsim
#> 33      Vmaxlen                             numeric vector length nsim
#> 34          LR5                             numeric vector length nsim
#> 35          LFR                             numeric vector length nsim
#> 36      Rmaxlen                             numeric vector length nsim
#> 37           DR                             numeric vector length nsim
#> 38         Cobs                             numeric vector length nsim
#> 39    CAA_nsamp                             numeric vector length nsim
#> 40      CAA_ESS                             numeric vector length nsim
#> 41    CAL_nsamp                             numeric vector length nsim
#> 42      CAL_ESS                             numeric vector length nsim
#> 43         Iobs                             numeric vector length nsim
#> 44        Btobs                             numeric vector length nsim
#> 45     Btbiascv                             numeric vector length nsim
#> 46         beta                             numeric vector length nsim
#> 47         Dobs                             numeric vector length nsim
#> 48    Recbiascv                             numeric vector length nsim
#> 49      TACFrac                             numeric vector length nsim
#> 50        TACSD                             numeric vector length nsim
#> 51      TAEFrac                             numeric vector length nsim
#> 52        TAESD                             numeric vector length nsim
#> 53  SizeLimFrac                             numeric vector length nsim
#> 54    SizeLimSD                             numeric vector length nsim
#> 55        Cbias                             numeric vector length nsim
#> 56     CAL_bins                                        numeric vector 
#> 57  CAL_binsmid                      numeric vector length(CAL_bins)-1
#> 58          L95                             numeric vector length nsim
#> 59       Perr_y numeric matrix dim = c(nsim, maxage+proyears+nyears-1)
#> 60  M_at_Length                   numeric matrix dim = c(n.lengths, 3)
#> 61        Asize                    numeric matrix dim = c(nsim, narea)
#> 62       Karray          numeric matrix dim = c(nsim, nyears+proyears)
#> 63    Linfarray          numeric matrix dim = c(nsim, nyears+proyears)
#> 64       Marray          numeric matrix dim = c(nsim, nyears+proyears)
#> 65        Krand          numeric matrix dim = c(nsim, nyears+proyears)
#> 66     Linfrand          numeric matrix dim = c(nsim, nyears+proyears)
#> 67        Mrand          numeric matrix dim = c(nsim, nyears+proyears)
#> 68         ageM          numeric matrix dim = c(nsim, nyears+proyears)
#> 69        age95          numeric matrix dim = c(nsim, nyears+proyears)
#> 70   M_ageArray   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 71      Mat_age   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 72       LatASD   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 73       Wt_age   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 74      Len_age   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 75          mov      numeric array dim = c(nsim, maxage, narea, narea)
#> 76         Find                   numeric matrix dim = c(nsim, nyears)
#> 77      dFfinal                             numeric vector length nsim
#> 78            V   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 79         retA   numeric array dim = c(nsim, maxage, nyears+proyears)
#> 80         retL numeric array dim = c(nsim, nCALbins, nyears+proyears)
#> 81     lenMbias                             numeric vector length nsim
#> 82        Mbias                             numeric vector length nsim
#> 83        Kbias                             numeric vector length nsim
#> 84       t0bias                             numeric vector length nsim
#> 85     Linfbias                             numeric vector length nsim
#> 86      LFCbias                             numeric vector length nsim
#> 87     FMSYbias                             numeric vector length nsim
#> 88   FMSY_Mbias                             numeric vector length nsim
#> 89  BMSY_B0bias                             numeric vector length nsim
#> 90     Irefbias                             numeric vector length nsim
#> 91     Brefbias                             numeric vector length nsim
#> 92     Crefbias                             numeric vector length nsim
#> 93        Dbias                             numeric vector length nsim
#> 94        hbias                             numeric vector length nsim
#> 95         hsim                             numeric vector length nsim
#>                                                                                                    Desc.
#> 1                                                                                     Virgin recruitment
#> 2                                                                                      Natural mortality
#> 3                                                                             Lorenzen M-weight exponent
#> 4                                                                          Inter-annual variability in M
#> 5                                                                                          Gradient in M
#> 6                                                                                              Steepness
#> 7                                                                             Stock-Recruit Relationship
#> 8                                                                              Recruitment process error
#> 9                                                                           Recruitment auto-correlation
#> 10                                                      Period for cyclical recruitment pattern in years
#> 11                    Amplitude in deviation from long-term average recruitment during recruitment cycle
#> 12                                                                               von B asymptotic length
#> 13                                                                              von B growth coefficient
#> 14                                                                                              von B t0
#> 15                                                                          Variability in length-at-age
#> 16                                                                         Inter-annual variability in K
#> 17                                                                                        Gradient in K 
#> 18                                                                      Inter-annual variability in Linf
#> 19                                                                                     Gradient in Linf 
#> 20                                                                                Length at 50% maturity
#> 21                                                                         Increment between L95 and L50
#> 22                                                                                             Depletion
#> 23                                                                     Size of Area 1 relative to Area 2
#> 24                                                                           Fraction of Stock in Area 1
#> 25                                                        Probablity of staying in Area 1 in 1 time-step
#> 26                                                                          Discard mortality (fraction)
#> 27                                                                                     Spatial targeting
#> 28                                                                         Inter-annual variability in F
#> 29                                                                              Gradient in catchability
#> 30                                                              Inter-annual variability in catchability
#> 31                                                  First length at 5% selection (isRel must be 'FALSE')
#> 32                                                First length at full selection (isRel must be 'FALSE')
#> 33                                            Vulnerability at asymptotic length (isRel must be 'FALSE')
#> 34                                                  First length at 5% retention (isRel must be 'FALSE')
#> 35                                                First length at full retention (isRel must be 'FALSE')
#> 36                                                Retention at asymptotic length (isRel must be 'FALSE')
#> 37                                                                                         Discard ratio
#> 38                                                               sampled catch obs error (log-normal sd)
#> 39                                                                                       CAA sample size
#> 40                                                                             CAA effective sample size
#> 41                                                                                       CAL sample size
#> 42                                                                             CAL effective sample size
#> 43                                                             Abundance index obs error (log normal sd)
#> 44                                                                        Observation error on Abundance
#> 45                             Uniform-log bounds for sampling persistent bias in current stock biomass.
#> 46                                                   Parameter controlling hyperstability/hyperdepletion
#> 47 Log-normal coefficient of variation controlling error in observations of stock depletion among years.
#> 48       Log-normal coefficient of variation for sampling persistent bias in recent recruitment strength
#> 49                                                                                      TAC imp fraction
#> 50                                                                                   TAC imp variability
#> 51                                                                                      TAE imp fraction
#> 52                                                                                   TAE imp variability
#> 53                                                                                     Size imp fraction
#> 54                                                                                  Size imp variability
#> 55                                                                                Bias in observed catch
#> 56                                                                      CAL bins. Largest must be > Linf
#> 57                                                                                Mid-points of CAL bins
#> 58                                                                                Length at 95% maturity
#> 59                   Recruitment process error by simulation and year (plus maxage for initial age comp)
#> 60        Data frame with column 1 lengths in absolute units, and colums 2:3 lower and upper bounds on M
#> 61                                                                           Size of areas by simulation
#> 62                                               von B K by simulation and year. No variability is added
#> 63                                            von B Linf by simulation and year. No variability is added
#> 64                                                     M by simulation and year. No variability is added
#> 65                                                  Inter-annual variability in K by simulation and year
#> 66                                               Inter-annual variability in Linf by simulation and year
#> 67                                                  Inter-annual variability in M by simulation and year
#> 68                                                            Age at 50% maturity by simulation and year
#> 69                                                            Age at 95% maturity by simulation and year
#> 70                                                M by simulation, age and year. No variability is added
#> 71                                                                 Maturity by simulation, age and year.
#> 72                                                                   standard deviation of length-at-age
#> 73                                                            Weight-at-age by simulation, age, and year
#> 74                                                            Length-at-age by simulation, age, and year
#> 75                                            Movement matrix by simulation, age, and area (to and from)
#> 76                                                 Historical normalized F/Effort by simulation and year
#> 77                                                                   Final gradient in fishing mortality
#> 78                                                              Selectivity by simulation, age, and year
#> 79                                                                Retention by simulation, age, and year
#> 80                                                             Retention by simulation, length, and year
#> 81                                                                   Bias in observed length-at-maturity
#> 82                                                                                    Bias in observed M
#> 83                                                                                    Bias in observed K
#> 84                                                                                   Bias in observed t0
#> 85                                                                                 Bias in observed Linf
#> 86                                                                                  Bias in observed LFC
#> 87                                                                                 Bias in observed FMSY
#> 88                                                                               Bias in observed FMSY_M
#> 89                                                                              Bias in observed BMSY_B0
#> 90                                                                                 Bias in observed Iref
#> 91                                                                                 Bias in observed Bref
#> 92                                                                                 Bias in observed Cref
#> 93                                                                                    Bias in observed D
#> 94                                                                                    Bias in observed h
#> 95                                                       Actual h values to pass directly to Data object
#>        Type
#> 1     Stock
#> 2     Stock
#> 3     Stock
#> 4     Stock
#> 5     Stock
#> 6     Stock
#> 7     Stock
#> 8     Stock
#> 9     Stock
#> 10    Stock
#> 11    Stock
#> 12    Stock
#> 13    Stock
#> 14    Stock
#> 15    Stock
#> 16    Stock
#> 17    Stock
#> 18    Stock
#> 19    Stock
#> 20    Stock
#> 21    Stock
#> 22    Stock
#> 23    Stock
#> 24    Stock
#> 25    Stock
#> 26    Stock
#> 27    Fleet
#> 28    Fleet
#> 29    Fleet
#> 30    Fleet
#> 31    Fleet
#> 32    Fleet
#> 33    Fleet
#> 34    Fleet
#> 35    Fleet
#> 36    Fleet
#> 37    Fleet
#> 38      Obs
#> 39      Obs
#> 40      Obs
#> 41      Obs
#> 42      Obs
#> 43      Obs
#> 44      Obs
#> 45      Obs
#> 46      Obs
#> 47      Obs
#> 48      Obs
#> 49      Imp
#> 50      Imp
#> 51      Imp
#> 52      Imp
#> 53      Imp
#> 54      Imp
#> 55 internal
#> 56 internal
#> 57 internal
#> 58 internal
#> 59 internal
#> 60 internal
#> 61 internal
#> 62 internal
#> 63 internal
#> 64 internal
#> 65 internal
#> 66 internal
#> 67 internal
#> 68 internal
#> 69 internal
#> 70 internal
#> 71 internal
#> 72 internal
#> 73 internal
#> 74 internal
#> 75 internal
#> 76 internal
#> 77 internal
#> 78 internal
#> 79 internal
#> 80 internal
#> 81 internal
#> 82 internal
#> 83 internal
#> 84 internal
#> 85 internal
#> 86 internal
#> 87 internal
#> 88 internal
#> 89 internal
#> 90 internal
#> 91 internal
#> 92 internal
#> 93 internal
#> 94 internal
#> 95 internal

validcpars("Obs", FALSE) # invalid Obs cpars

#>             Var.              Desc. Type
#> 1        Cbiascv     Not stochastic  Obs
#> 2        Ibiascv Not currently used  Obs
#> 3     LenMbiascv       Use lenMbias  Obs
#> 4        Mbiascv          Use Mbias  Obs
#> 5        Kbiascv          Use Kbias  Obs
#> 6       t0biascv         Use t0bias  Obs
#> 7     Linfbiascv       Use Linfbias  Obs
#> 8      LFCbiascv        Use LFCbias  Obs
#> 9      LFSbiascv        Use LFSbias  Obs
#> 10    FMSYbiascv       Use FMSYbias  Obs
#> 11  FMSY_Mbiascv     Use FMSY_Mbias  Obs
#> 12 BMSY_B0biascv    Use BMSY_B0bias  Obs
#> 13    Irefbiascv       Use Irefbias  Obs
#> 14    Brefbiascv       Use Brefbias  Obs
#> 15    Crefbiascv       Use Crefbias  Obs
#> 16       Dbiascv          Use Dbias  Obs
#> 17       hbiascv          Use hbias  Obs